Prefetch cell cycle genes

Based on the human cell cycle genes, the cell cycle genes of the corresponding species were captured by homologous gene conversion.

Usage

CycGenePrefetch(
  species = "Homo_sapiens",
  Ensembl_version = NULL,
  mirror = NULL,
  max_tries = 5,
  use_cached_gene = TRUE,
  verbose = TRUE
)

Arguments

species

Latin names for animals, i.e., "Homo_sapiens", "Mus_musculus"

Ensembl_version

Ensembl version. NULL uses the latest.

mirror

Specify an Ensembl mirror to connect to. The valid options here are "www", "uswest", "useast", "asia".

max_tries

The maximum number of attempts to connect with the BioMart service.

use_cached_gene

Whether to use previously cached cell cycle gene conversion results for the species.

verbose

Whether to print the message. Default is TRUE.

Value

A list of S-phase and G2M-phase genes.

See also

Examples

ccgenes <- CycGenePrefetch("Homo_sapiens")
#> ℹ [2026-09-03 03:18:40] Prefetching cell cycle genes for "Homo_sapiens" ...
#> ✔ [2026-09-03 03:18:40] Cell cycle gene prefetching completed "Homo_sapiens"
str(ccgenes)
#> List of 5
#>  $ res         : NULL
#>  $ S           : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ G2M         : chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...
#>  $ cc_S_genes  : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ cc_G2M_genes: chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...

ccgenes <- CycGenePrefetch("Mus_musculus")
#> ℹ [2026-09-03 03:18:40] Prefetching cell cycle genes for "Mus_musculus" ...
#> ℹ [2026-09-03 03:18:41] Connect to the Ensembl archives...
#> ℹ [2026-09-03 03:18:41] Using the 116 version of ensembl database...
#> ℹ [2026-09-03 03:18:41] Downloading the ensembl database from https://jun2026.archive.ensembl.org...
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying useast mirror
#> ! [2026-09-03 03:18:45] <simpleError in .chooseEnsemblMirror(mirror = mirror, http_config = http_config): Unable to query any Ensembl site>
#> ! [2026-09-03 03:18:45] Get errors when connecting with ensembl database...
#> ! [2026-09-03 03:18:46] Retrying...
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying asia mirror
#> ! [2026-09-03 03:18:50] <simpleError in .chooseEnsemblMirror(mirror = mirror, http_config = http_config): Unable to query any Ensembl site>
#> ! [2026-09-03 03:18:50] Get errors when connecting with ensembl database...
#> ! [2026-09-03 03:18:51] Retrying...
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying asia mirror
#> ! [2026-09-03 03:18:55] <simpleError in .chooseEnsemblMirror(mirror = mirror, http_config = http_config): Unable to query any Ensembl site>
#> ! [2026-09-03 03:18:55] Get errors when connecting with ensembl database...
#> ! [2026-09-03 03:18:56] Retrying...
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying useast mirror
#> ! [2026-09-03 03:19:00] <simpleError in .chooseEnsemblMirror(mirror = mirror, http_config = http_config): Unable to query any Ensembl site>
#> ! [2026-09-03 03:19:00] Get errors when connecting with ensembl database...
#> ! [2026-09-03 03:19:01] Retrying...
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying asia mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying www mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying useast mirror
#> Ensembl site unresponsive, trying asia mirror
#> ! [2026-09-03 03:19:05] <simpleError in .chooseEnsemblMirror(mirror = mirror, http_config = http_config): Unable to query any Ensembl site>
#> ! [2026-09-03 03:19:05] Get errors when connecting with ensembl database...
#> Error in try_get(expr = {    if (isTRUE(use_current_release)) {        mirror_candidates <- unique(stats::na.omit(c(mirror,             "useast", "uswest", "asia", "www")))        mart_try <- NULL        for (mirror_i in mirror_candidates) {            mart_try <- tryCatch(biomaRt::useEnsembl(biomart = "ensembl",                 mirror = mirror_i), error = function(e) e)            if (!inherits(mart_try, "error")) {                break            }        }        if (inherits(mart_try, "error")) {            stop(mart_try)        }        mart_try    }    else if (!is.null(mirror)) {        biomaRt::useEnsembl(biomart = "ensembl", mirror = mirror)    }    else {        mart_try <- tryCatch(biomaRt::useMart(biomart = "ensembl",             host = url), error = function(e) e)        if (inherits(mart_try, "error")) {            mirror_candidates <- c("useast", "uswest", "asia",                 "www")            for (mirror_i in mirror_candidates) {                mart_mirror <- tryCatch(biomaRt::useEnsembl(biomart = "ensembl",                   mirror = mirror_i), error = function(e) e)                if (!inherits(mart_mirror, "error")) {                  mart_try <- mart_mirror                  break                }            }        }        if (inherits(mart_try, "error")) {            stop(mart_try)        }        mart_try    }}, max_tries = max_tries, error_message = "Get errors when connecting with ensembl database..."): <simpleError in .chooseEnsemblMirror(mirror = mirror, http_config =
#> http_config): Unable to query any Ensembl site>
str(ccgenes)
#> List of 5
#>  $ res         : NULL
#>  $ S           : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ G2M         : chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...
#>  $ cc_S_genes  : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ cc_G2M_genes: chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...