Get expression data from Assay5 or Seurat object

A re-implementation of the SeuratObject::GetAssayData function to compatible with Assay5 objects.

Usage

GetAssayData5(object, ...)

# S3 method for class 'Seurat'
GetAssayData5(object, layer = "counts", assay = NULL, verbose = TRUE, ...)

# S3 method for class 'Assay5'
GetAssayData5(
  object,
  layer = "counts",
  join_layers = TRUE,
  verbose = TRUE,
  ...
)

# S3 method for class 'Assay'
GetAssayData5(object, layer = "counts", verbose = TRUE, ...)

Arguments

object

An object

...

Additional arguments passed to SeuratObject::GetAssayData.

layer

Name of layer to get or set

assay

Specific assay to get data from or set data for; defaults to the default assay

verbose

Whether to print messages. Default is TRUE.

join_layers

Whether to join layers if the object is an Assay5 object. Default is `TRUE`.

Value

A matrix or data frame containing the assay data.

See also

Examples

data(pancreas_sub)
GetAssayData5(
  pancreas_sub,
  layer = "counts",
  assay = "RNA"
)[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>               AAACCTGAGCCTTGAT AAACCTGAGGCAATTA AAACCTGTCTTTCCTC
#> Xkr4                         .                .                .
#> Mrpl15                       .                .                .
#> 4732440D04Rik                .                .                .
#> Gm26901                      1                .                .
#> Sntg1                        .                .                .
#>               AAACGGGAGATATGGT AAACGGGCAAGGACAC
#> Xkr4                         .                .
#> Mrpl15                       .                .
#> 4732440D04Rik                .                .
#> Gm26901                      .                .
#> Sntg1                        .                .

data(panc8_sub)
GetAssayData5(
  panc8_sub,
  layer = "counts",
  assay = "RNA"
)[1:5, 1:5]
#> 5 x 5 sparse Matrix of class "dgCMatrix"
#>        D101_17 D101_22 D101_31 D101_43 D101_51
#> A1CF         .       .       .       .       .
#> A2M          .       .       .       .       .
#> A4GALT       .       .       .       .       .
#> AAAS         .       .       .       .       .
#> AACS         .       .       .       .       .