List cached databases

Retrieves information about databases based on a given species and database name.

Usage

ListDB(species = c("Homo_sapiens", "Mus_musculus"), db = NULL)

Arguments

species

A character vector of species for which to retrieve database information. Default is c("Homo_sapiens", "Mus_musculus").

db

The pattern to match against the database names. Default is NULL, which matches all databases.

Value

A data frame containing information about the databases, including a Species column and a DB column.

See also

Examples

ListDB(species = "Homo_sapiens")
#>  [1] identifier version    comment    timestamp  date       db_version
#>  [7] db_name    file       Species    DB        
#> <0 rows> (or 0-length row.names)
ListDB(species = c("Homo_sapiens", "Mus_musculus"))
#>                                                         identifier version
#> 1 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#> 2 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#> 3 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#> 4 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#> 5 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#> 6 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#>                                  comment  timestamp                       date
#> 1       9.0.0 nterm:6595|Mus_musculus-DO 1786063506 2026-08-07 00:45:05.953748
#> 2 2026-08-07 nterm:10898|Mus_musculus-MP 1786063502 2026-08-07 00:45:01.685495
#> 3   3.23.0 nterm:2065|Mus_musculus-GO_CC 1786063452 2026-08-07 00:44:12.034893
#> 4  3.23.0 nterm:14957|Mus_musculus-GO_BP 1786063451 2026-08-07 00:44:10.804991
#> 5         CSPA nterm:1|Mus_musculus-CSPA 1786063362 2026-08-07 00:42:42.105837
#> 6    AnimalTFDB4 nterm:2|Mus_musculus-TF 1786062046 2026-08-07 00:20:46.133101
#>               db_version            db_name
#> 1       9.0.0 nterm:6595    Mus_musculus-DO
#> 2 2026-08-07 nterm:10898    Mus_musculus-MP
#> 3      3.23.0 nterm:2065 Mus_musculus-GO_CC
#> 4     3.23.0 nterm:14957 Mus_musculus-GO_BP
#> 5           CSPA nterm:1  Mus_musculus-CSPA
#> 6    AnimalTFDB4 nterm:2    Mus_musculus-TF
#>                                                                    file
#> 1 /home/runner/.cache/R/R.cache/48fdd5e44751c0e22be021269c96d88c.Rcache
#> 2 /home/runner/.cache/R/R.cache/6574de2ae930e2958426177617ed245a.Rcache
#> 3 /home/runner/.cache/R/R.cache/fb4602feef7bb4dc9c034fd6849e3dc5.Rcache
#> 4 /home/runner/.cache/R/R.cache/a759db6d549ed9ffa83592264ac75787.Rcache
#> 5 /home/runner/.cache/R/R.cache/267624ab3d48b245ce01deac5165a7b1.Rcache
#> 6 /home/runner/.cache/R/R.cache/82f852e1d9ed5c0a5ad672b843c18742.Rcache
#>        Species    DB
#> 1 Mus_musculus    DO
#> 2 Mus_musculus    MP
#> 3 Mus_musculus GO_CC
#> 4 Mus_musculus GO_BP
#> 5 Mus_musculus  CSPA
#> 6 Mus_musculus    TF
ListDB(species = "Mus_musculus", db = "GO_BP")
#>                                                         identifier version
#> 1 Rcache v0.1.7 (R package R.cache by Henrik Bengtsson)              0.1.7
#>                                 comment  timestamp                       date
#> 1 3.23.0 nterm:14957|Mus_musculus-GO_BP 1786063451 2026-08-07 00:44:10.804991
#>           db_version            db_name
#> 1 3.23.0 nterm:14957 Mus_musculus-GO_BP
#>                                                                    file
#> 1 /home/runner/.cache/R/R.cache/a759db6d549ed9ffa83592264ac75787.Rcache
#>        Species    DB
#> 1 Mus_musculus GO_BP