Run WOT analysis

Run WOT analysis

RunWOT(
  srt = NULL,
  assay_X = "RNA",
  slot_X = "counts",
  assay_layers = c("spliced", "unspliced"),
  slot_layers = "counts",
  adata = NULL,
  group_by = NULL,
  time_field = "Time",
  growth_iters = 3L,
  tmap_out = "tmaps/tmap_out",
  time_from = NULL,
  time_to = NULL,
  get_coupling = FALSE,
  recalculate = FALSE,
  palette = "Paired",
  palcolor = NULL,
  show_plot = TRUE,
  save = FALSE,
  dpi = 300,
  dirpath = "./",
  fileprefix = "",
  return_seurat = !is.null(srt)
)

Arguments

srt

A Seurat object.

assay_X

Assay to convert as the main data matrix (X) in the anndata object.

slot_X

Slot name for assay_X in the Seurat object.

assay_layers

Assays to convert as layers in the anndata object.

slot_layers

Slot names for the assay_layers in the Seurat object.

adata

An anndata object.

group_by

Variable to use for grouping cells in the Seurat object.

time_field

A character string specifying the column name in `adata.obs` or `srt@meta.data` that contains the time information.

growth_iters

An integer specifying the number of growth iterations to perform during the OT Model computation. Default is 3.

tmap_out

A character string specifying the path to store the computed transport maps.

time_from

A numeric value specifying the starting time point for trajectory and fate analysis.

time_to

A numeric value specifying the ending time point for trajectory and fate analysis. If not provided, only trajectory and fate analysis for the specified `time_from` will be performed.

get_coupling

A logical value indicating whether to compute and store the coupling matrix between the specified `time_from` and `time_to`. Default is FALSE.

recalculate

A logical value indicating whether to recalculate the transport maps even if they already exist at the specified `tmap_out` location. Default is FALSE.

palette

The palette to use for coloring cells.

palcolor

A vector of colors to use as the palette.

show_plot

Whether to show the PAGA plot.

save

Whether to save the PAGA plots.

dpi

The DPI (dots per inch) for saving the PAGA plot.

dirpath

The directory to save the PAGA plots.

fileprefix

The file prefix to use for the PAGA plots.

return_seurat

Whether to return a Seurat object instead of an anndata object. Default is TRUE.

See also