Convert an anndata object to a seurat object using reticulate

Convert an anndata object to a seurat object using reticulate

Usage

adata_to_srt(adata, verbose = TRUE)

Arguments

adata

A connected python anndata object.

verbose

Whether to print messages.

Examples

data(pancreas_sub)
adata <- srt_to_adata(pancreas_sub)
#> ℹ [2025-09-03 14:50:17] Checking 2 packages in environment: scop_env
#> ℹ [2025-09-03 14:50:18] Retrieving package list for environment: scop_env
#> ℹ [2025-09-03 14:50:19] Found 175 packages installed
#> ✔ [2025-09-03 14:50:19] scanpy version: 1.11.4
#> ✔ [2025-09-03 14:50:19] numpy version: 1.26.4
#> ℹ [2025-09-03 14:50:20] Converting <Seurat> object to <AnnData> object...
#> Error: Required version of NumPy not available: installation of Numpy >= 1.6 not found
adata <- RunPAGA(
  adata = adata,
  group_by = "SubCellType",
  linear_reduction = "PCA",
  nonlinear_reduction = "UMAP"
)
#> Error: object 'adata' not found
srt <- adata_to_srt(adata)
#> Error: object 'adata' not found
srt
#> Error: object 'srt' not found

# Or convert a h5ad file to Seurat object
library(reticulate)
check_python("scanpy")
#> ℹ [2025-09-03 14:50:22] Checking 1 packages in environment: scop_env
#> ℹ [2025-09-03 14:50:22] Retrieving package list for environment: scop_env
#> ℹ [2025-09-03 14:50:24] Found 175 packages installed
#> ✔ [2025-09-03 14:50:24] scanpy version: 1.11.4
sc <- import("scanpy")
adata <- sc$read_h5ad("pancreas.h5ad")
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named): FileNotFoundError: [Errno 2] Unable to synchronously open file (unable to open file: name = 'pancreas.h5ad', errno = 2, error message = 'No such file or directory', flags = 0, o_flags = 0)
#> Run `reticulate::py_last_error()` for details.
srt <- adata_to_srt(adata)
#> Error: object 'adata' not found
srt
#> Error: object 'srt' not found