Annotate Features

Annotate features in a Seurat object with additional metadata from databases or a GTF file.

Usage

AnnotateFeatures(
  srt,
  species = "Homo_sapiens",
  IDtype = c("symbol", "ensembl_id", "entrez_id"),
  db = NULL,
  db_update = FALSE,
  db_version = "latest",
  convert_species = TRUE,
  Ensembl_version = 103,
  mirror = NULL,
  gtf = NULL,
  merge_gtf_by = "gene_name",
  columns = c("seqname", "feature", "start", "end", "strand", "gene_id", "gene_name",
    "gene_type"),
  assays = "RNA",
  overwrite = FALSE
)

Arguments

srt

Seurat object to be annotated.

species

Name of the species to be used for annotation. Default is "Homo_sapiens".

IDtype

Type of identifier to use for annotation. Default is "symbol" with options "symbol", "ensembl_id", and "entrez_id".

db

Vector of database names to be used for annotation. Default is NULL.

db_update

Logical value indicating whether to update the database. Default is FALSE.

db_version

Version of the database to use. Default is "latest".

convert_species

A logical value indicating whether to use a species-converted database when the annotation is missing for the specified species. The default value is TRUE.

Ensembl_version

Version of the Ensembl database to use. Default is 103.

mirror

URL of the mirror to use for Ensembl database. Default is NULL.

gtf

Path to the GTF file to be used for annotation. Default is NULL.

merge_gtf_by

Column name to merge the GTF file by. Default is "gene_name".

columns

Vector of column names to be used from the GTF file. Default is "seqname", "feature", "start", "end", "strand", "gene_id", "gene_name", "gene_type".

assays

Character vector of assay names to be annotated. Default is "RNA".

overwrite

Logical value indicating whether to overwrite existing metadata. Default is FALSE.

See also

Examples

data(pancreas_sub)
pancreas_sub <- AnnotateFeatures(
  pancreas_sub,
  species = "Mus_musculus",
  db = c(
    "Chromosome",
    "GeneType",
    "Enzyme",
    "TF",
    "CSPA",
    "VerSeDa"
  )
)
#> ℹ [2025-09-09 13:49:33] Species: Mus_musculus
#> ℹ [2025-09-09 13:49:33] Installing package: org.Mm.eg.db...
#> ℹ Loading metadata database
#> ✔ Loading metadata database ... done
#> 
#>  
#> → Will install 37 packages.
#> → All 37 packages (0 B) are cached.
#> + AnnotationDbi      1.70.0  [bld]
#> + Biobase            2.68.0  [bld][cmp]
#> + BiocGenerics       0.54.0  [bld]
#> + Biostrings         2.76.0  [bld][cmp]
#> + DBI                1.2.3   
#> + GenomeInfoDb       1.44.2  [bld]
#> + GenomeInfoDbData   1.2.14  [bld]
#> + IRanges            2.42.0  [bld][cmp]
#> + KEGGREST           1.48.1  [bld]
#> + R6                 2.6.1   
#> + RSQLite            2.4.3   
#> + S4Vectors          0.46.0  [bld][cmp]
#> + UCSC.utils         1.4.0   [bld]
#> + XVector            0.48.0  [bld][cmp]
#> + askpass            1.2.1   
#> + bit                4.6.0   
#> + bit64              4.6.0-1 
#> + blob               1.2.4   
#> + cachem             1.1.0   
#> + cli                3.6.5   
#> + crayon             1.5.3   
#> + curl               7.0.0    + ✔ libcurl4-openssl-dev, ✔ libssl-dev
#> + fastmap            1.2.0   
#> + generics           0.1.4   
#> + glue               1.8.0   
#> + httr               1.4.7   
#> + jsonlite           2.0.0   
#> + lifecycle          1.0.4   
#> + memoise            2.0.1   
#> + mime               0.13    
#> + openssl            2.3.3    + ✔ libssl-dev
#> + org.Mm.eg.db       3.21.0  [bld]
#> + pkgconfig          2.0.3   
#> + png                0.1-8    + ✔ libpng-dev
#> + rlang              1.1.6   
#> + sys                3.4.3   
#> + vctrs              0.6.5   
#> ✔ All system requirements are already installed.
#>   
#> ℹ No downloads are needed, 37 pkgs are cached
#> ℹ Installing system requirements
#> ℹ Executing `sudo sh -c apt-get -y update`
#> Get:1 file:/etc/apt/apt-mirrors.txt Mirrorlist [144 B]
#> Hit:2 http://azure.archive.ubuntu.com/ubuntu noble InRelease
#> Get:3 http://azure.archive.ubuntu.com/ubuntu noble-updates InRelease [126 kB]
#> Get:4 http://azure.archive.ubuntu.com/ubuntu noble-backports InRelease [126 kB]
#> Get:5 http://azure.archive.ubuntu.com/ubuntu noble-security InRelease [126 kB]
#> Hit:6 https://packages.microsoft.com/repos/azure-cli noble InRelease
#> Get:7 https://packages.microsoft.com/ubuntu/24.04/prod noble InRelease [3600 B]
#> Get:8 http://azure.archive.ubuntu.com/ubuntu noble-updates/main amd64 Packages [1390 kB]
#> Get:9 http://azure.archive.ubuntu.com/ubuntu noble-updates/main amd64 Components [175 kB]
#> Get:10 http://azure.archive.ubuntu.com/ubuntu noble-updates/universe amd64 Packages [1481 kB]
#> Get:11 http://azure.archive.ubuntu.com/ubuntu noble-updates/universe amd64 Components [377 kB]
#> Get:12 http://azure.archive.ubuntu.com/ubuntu noble-updates/restricted amd64 Components [212 B]
#> Get:13 http://azure.archive.ubuntu.com/ubuntu noble-updates/multiverse amd64 Components [940 B]
#> Get:14 http://azure.archive.ubuntu.com/ubuntu noble-backports/main amd64 Components [7064 B]
#> Get:15 http://azure.archive.ubuntu.com/ubuntu noble-backports/universe amd64 Components [19.2 kB]
#> Get:16 http://azure.archive.ubuntu.com/ubuntu noble-backports/restricted amd64 Components [216 B]
#> Get:17 http://azure.archive.ubuntu.com/ubuntu noble-backports/multiverse amd64 Components [212 B]
#> Get:18 http://azure.archive.ubuntu.com/ubuntu noble-security/main amd64 Components [21.6 kB]
#> Get:19 http://azure.archive.ubuntu.com/ubuntu noble-security/universe amd64 Components [52.2 kB]
#> Get:20 http://azure.archive.ubuntu.com/ubuntu noble-security/restricted amd64 Components [212 B]
#> Get:21 http://azure.archive.ubuntu.com/ubuntu noble-security/multiverse amd64 Components [212 B]
#> Get:22 https://packages.microsoft.com/ubuntu/24.04/prod noble/main amd64 Packages [51.4 kB]
#> Get:23 https://packages.microsoft.com/ubuntu/24.04/prod noble/main arm64 Packages [35.3 kB]
#> Fetched 3994 kB in 1s (6506 kB/s)
#> Reading package lists...
#> ℹ Executing `sudo sh -c apt-get -y install libcurl4-openssl-dev libssl-dev libpng-dev`
#> Reading package lists...
#> Building dependency tree...
#> Reading state information...
#> libcurl4-openssl-dev is already the newest version (8.5.0-2ubuntu10.6).
#> libssl-dev is already the newest version (3.0.13-0ubuntu3.5).
#> libpng-dev is already the newest version (1.6.43-5build1).
#> 0 upgraded, 0 newly installed, 0 to remove and 20 not upgraded.
#> ℹ Building GenomeInfoDbData 1.2.14
#> ✔ Installed askpass 1.2.1  (100ms)
#> ✔ Installed bit 4.6.0  (160ms)
#> ✔ Installed bit64 4.6.0-1  (184ms)
#> ✔ Installed blob 1.2.4  (66ms)
#> ✔ Installed cachem 1.1.0  (66ms)
#> ✔ Installed cli 3.6.5  (64ms)
#> ✔ Installed crayon 1.5.3  (64ms)
#> ✔ Installed curl 7.0.0  (67ms)
#> ✔ Installed DBI 1.2.3  (97ms)
#> ✔ Installed fastmap 1.2.0  (72ms)
#> ✔ Installed generics 0.1.4  (91ms)
#> ℹ Building BiocGenerics 0.54.0
#> ✔ Installed glue 1.8.0  (87ms)
#> ✔ Installed httr 1.4.7  (77ms)
#> ✔ Installed jsonlite 2.0.0  (100ms)
#> ✔ Installed lifecycle 1.0.4  (97ms)
#> ✔ Installed memoise 2.0.1  (92ms)
#> ✔ Installed mime 0.13  (115ms)
#> ✔ Installed openssl 2.3.3  (124ms)
#> ✔ Installed pkgconfig 2.0.3  (132ms)
#> ✔ Installed png 0.1-8  (86ms)
#> ✔ Installed R6 2.6.1  (87ms)
#> ✔ Installed rlang 1.1.6  (96ms)
#> ✔ Installed RSQLite 2.4.3  (104ms)
#> ✔ Installed sys 3.4.3  (146ms)
#> ✔ Installed vctrs 0.6.5  (117ms)
#> ✔ Built GenomeInfoDbData 1.2.14 (1.9s)
#> ✔ Installed GenomeInfoDbData 1.2.14  (1.1s)
#> ✔ Built BiocGenerics 0.54.0 (3.3s)
#> ✔ Installed BiocGenerics 0.54.0  (1s)
#> ℹ Building Biobase 2.68.0
#> ℹ Building S4Vectors 0.46.0
#> ✔ Built Biobase 2.68.0 (6.5s)
#> ✔ Installed Biobase 2.68.0  (58ms)
#> ✔ Built S4Vectors 0.46.0 (15.9s)
#> ✔ Installed S4Vectors 0.46.0  (53ms)
#> ℹ Building IRanges 2.42.0
#> ℹ Building UCSC.utils 1.4.0
#> ✔ Built UCSC.utils 1.4.0 (2.5s)
#> ✔ Installed UCSC.utils 1.4.0  (1s)
#> ✔ Built IRanges 2.42.0 (37.4s)
#> ✔ Installed IRanges 2.42.0  (1s)
#> ℹ Building GenomeInfoDb 1.44.2
#> ℹ Building XVector 0.48.0
#> ✔ Built GenomeInfoDb 1.44.2 (7.4s)
#> ✔ Installed GenomeInfoDb 1.44.2  (1.1s)
#> ✔ Built XVector 0.48.0 (9s)
#> ✔ Installed XVector 0.48.0  (1s)
#> ℹ Building Biostrings 2.76.0
#> ✔ Built Biostrings 2.76.0 (19.1s)
#> ✔ Installed Biostrings 2.76.0  (1.2s)
#> ℹ Building KEGGREST 1.48.1
#> ✔ Built KEGGREST 1.48.1 (4.8s)
#> ✔ Installed KEGGREST 1.48.1  (1s)
#> ℹ Building AnnotationDbi 1.70.0
#> ✔ Built AnnotationDbi 1.70.0 (12s)
#> ✔ Installed AnnotationDbi 1.70.0  (66ms)
#> ℹ Building org.Mm.eg.db 3.21.0
#> ✔ Built org.Mm.eg.db 3.21.0 (3m 12.8s)
#> ✔ Installed org.Mm.eg.db 3.21.0  (1.8s)
#> ✔ 1 pkg + 36 deps: added 37 [5m 19.2s]
#> ℹ [2025-09-09 13:54:52] Installing package: GO.db...
#>  
#> → Will install 1 package.
#> → The package (0 B) is cached.
#> + GO.db   3.21.0 [bld]
#> ✔ All system requirements are already installed.
#>   
#> ℹ No downloads are needed, 1 pkg is cached
#> ℹ Installing system requirements
#> ℹ Executing `sudo sh -c apt-get -y update`
#> Get:1 file:/etc/apt/apt-mirrors.txt Mirrorlist [144 B]
#> Hit:6 https://packages.microsoft.com/repos/azure-cli noble InRelease
#> Hit:2 http://azure.archive.ubuntu.com/ubuntu noble InRelease
#> Hit:7 https://packages.microsoft.com/ubuntu/24.04/prod noble InRelease
#> Hit:3 http://azure.archive.ubuntu.com/ubuntu noble-updates InRelease
#> Hit:4 http://azure.archive.ubuntu.com/ubuntu noble-backports InRelease
#> Hit:5 http://azure.archive.ubuntu.com/ubuntu noble-security InRelease
#> Reading package lists...
#> ℹ Executing `sudo sh -c apt-get -y install libcurl4-openssl-dev libssl-dev libpng-dev`
#> Reading package lists...
#> Building dependency tree...
#> Reading state information...
#> libcurl4-openssl-dev is already the newest version (8.5.0-2ubuntu10.6).
#> libssl-dev is already the newest version (3.0.13-0ubuntu3.5).
#> libpng-dev is already the newest version (1.6.43-5build1).
#> 0 upgraded, 0 newly installed, 0 to remove and 20 not upgraded.
#> ℹ Building GO.db 3.21.0
#> ✔ Built GO.db 3.21.0 (36.5s)
#> ✔ Installed GO.db 3.21.0  (1.5s)
#> ✔ 1 pkg + 36 deps: kept 36, added 1 [41.2s]
#> ℹ [2025-09-09 13:55:34] Installing package: GOSemSim...
#>  
#> → Will install 9 packages.
#> → All 9 packages (0 B) are cached.
#> + GOSemSim      2.34.0 [bld][cmp]
#> + R.methodsS3   1.8.2  
#> + R.oo          1.27.1 
#> + R.utils       2.13.0 
#> + Rcpp          1.1.0  
#> + digest        0.6.37 
#> + fs            1.6.6   + ✔ make
#> + rappdirs      0.3.3  
#> + yulab.utils   0.2.1  
#> ✔ All system requirements are already installed.
#>   
#> ℹ No downloads are needed, 9 pkgs are cached
#> ℹ Installing system requirements
#> ℹ Executing `sudo sh -c apt-get -y update`
#> Get:1 file:/etc/apt/apt-mirrors.txt Mirrorlist [144 B]
#> Hit:2 http://azure.archive.ubuntu.com/ubuntu noble InRelease
#> Hit:3 http://azure.archive.ubuntu.com/ubuntu noble-updates InRelease
#> Hit:4 http://azure.archive.ubuntu.com/ubuntu noble-backports InRelease
#> Hit:5 http://azure.archive.ubuntu.com/ubuntu noble-security InRelease
#> Hit:6 https://packages.microsoft.com/repos/azure-cli noble InRelease
#> Hit:7 https://packages.microsoft.com/ubuntu/24.04/prod noble InRelease
#> Reading package lists...
#> ℹ Executing `sudo sh -c apt-get -y install make libcurl4-openssl-dev libssl-dev libpng-dev`
#> Reading package lists...
#> Building dependency tree...
#> Reading state information...
#> make is already the newest version (4.3-4.1build2).
#> libcurl4-openssl-dev is already the newest version (8.5.0-2ubuntu10.6).
#> libssl-dev is already the newest version (3.0.13-0ubuntu3.5).
#> libpng-dev is already the newest version (1.6.43-5build1).
#> 0 upgraded, 0 newly installed, 0 to remove and 20 not upgraded.
#> ✔ Installed digest 0.6.37  (64ms)
#> ✔ Installed fs 1.6.6  (80ms)
#> ✔ Installed R.methodsS3 1.8.2  (99ms)
#> ✔ Installed R.oo 1.27.1  (131ms)
#> ✔ Installed R.utils 2.13.0  (62ms)
#> ✔ Installed rappdirs 0.3.3  (103ms)
#> ✔ Installed Rcpp 1.1.0  (67ms)
#> ✔ Installed yulab.utils 0.2.1  (47ms)
#> ℹ Building GOSemSim 2.34.0
#> ✔ Built GOSemSim 2.34.0 (13.2s)
#> ✔ Installed GOSemSim 2.34.0  (1s)
#> ✔ 1 pkg + 45 deps: kept 37, added 9 [18.3s]
#> ℹ [2025-09-09 13:55:52] All packages installed successfully
#> ℹ [2025-09-09 13:55:53] Preparing database: Chromosome
#> ℹ [2025-09-09 13:55:53] Preparing database: GeneType
#> ℹ [2025-09-09 13:55:54] Preparing database: Enzyme
#> ℹ [2025-09-09 13:55:55] Preparing database: TF
#> ℹ [2025-09-09 13:55:57] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Mus_musculus_TF'
#> ! [2025-09-09 13:55:57] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Mus_musculus_TF>
#> ! [2025-09-09 13:55:57] Failed to download using "auto"
#> ℹ [2025-09-09 13:55:59] Error in utils::download.file(url = url, destfile = destfile, method = method, : 'wget' call had nonzero exit status
#> ! [2025-09-09 13:55:59] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Mus_musculus_TF>
#> ! [2025-09-09 13:55:59] Failed to download using "wget"
#> ℹ [2025-09-09 13:56:01] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Mus_musculus_TF'
#> ! [2025-09-09 13:56:01] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Mus_musculus_TF>
#> ! [2025-09-09 13:56:01] Failed to download using "libcurl"
#> Warning: incomplete final line found by readTableHeader on '/tmp/RtmpFh70D9/filea8ceb8d22e7'
#> ℹ [2025-09-09 13:56:05] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Mus_musculus_Cof'
#> ! [2025-09-09 13:56:05] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Mus_musculus_Cof>
#> ! [2025-09-09 13:56:05] Failed to download using "auto"
#> ℹ [2025-09-09 13:56:07] Error in utils::download.file(url = url, destfile = destfile, method = method, : 'wget' call had nonzero exit status
#> ! [2025-09-09 13:56:07] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Mus_musculus_Cof>
#> ! [2025-09-09 13:56:07] Failed to download using "wget"
#> ℹ [2025-09-09 13:56:09] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Mus_musculus_Cof'
#> ! [2025-09-09 13:56:09] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Mus_musculus_Cof>
#> ! [2025-09-09 13:56:09] Failed to download using "libcurl"
#> Warning: incomplete final line found by readTableHeader on '/tmp/RtmpFh70D9/filea8ceb8d22e7'
#> ! [2025-09-09 13:56:11] Use the human annotation to create the TF database for Mus_musculus
#> ℹ [2025-09-09 13:56:13] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Homo_sapiens_TF'
#> ! [2025-09-09 13:56:13] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Homo_sapiens_TF>
#> ! [2025-09-09 13:56:13] Failed to download using "auto"
#> ℹ [2025-09-09 13:56:15] Error in utils::download.file(url = url, destfile = destfile, method = method, : 'wget' call had nonzero exit status
#> ! [2025-09-09 13:56:15] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Homo_sapiens_TF>
#> ! [2025-09-09 13:56:15] Failed to download using "wget"
#> ℹ [2025-09-09 13:56:17] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Homo_sapiens_TF'
#> ! [2025-09-09 13:56:17] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/TF_list_final/Homo_sapiens_TF>
#> ! [2025-09-09 13:56:17] Failed to download using "libcurl"
#> Warning: incomplete final line found by readTableHeader on '/tmp/RtmpFh70D9/filea8ceb8d22e7'
#> ℹ [2025-09-09 13:56:20] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Homo_sapiens_Cof'
#> ! [2025-09-09 13:56:20] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Homo_sapiens_Cof>
#> ! [2025-09-09 13:56:20] Failed to download using "auto"
#> ℹ [2025-09-09 13:56:22] Error in utils::download.file(url = url, destfile = destfile, method = method, : 'wget' call had nonzero exit status
#> ! [2025-09-09 13:56:22] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Homo_sapiens_Cof>
#> ! [2025-09-09 13:56:22] Failed to download using "wget"
#> ℹ [2025-09-09 13:56:25] Error in utils::download.file(url = url, destfile = destfile, method = method, : cannot open URL 'https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Homo_sapiens_Cof'
#> ! [2025-09-09 13:56:25] Cannot download from the url: <https://guolab.wchscu.cn/AnimalTFDB4_static/download/Cof_list_final/Homo_sapiens_Cof>
#> ! [2025-09-09 13:56:25] Failed to download using "libcurl"
#> Warning: incomplete final line found by readTableHeader on '/tmp/RtmpFh70D9/filea8ceb8d22e7'
#> Error in data.frame(Term = "TF", symbol = tf[["Symbol"]]): arguments imply differing number of rows: 1, 0
head(
  GetFeaturesData(
    pancreas_sub,
    assays = "RNA"
  )
)
#> data frame with 0 columns and 6 rows

if (FALSE) { # \dontrun{
# Annotate features using a GTF file
pancreas_sub <- AnnotateFeatures(
  pancreas_sub,
  gtf = "/refdata-gex-mm10-2020-A/genes/genes.gtf"
)
head(
  GetFeaturesData(
    pancreas_sub,
    assays = "RNA"
  )
)
} # }