AnnotateFeatures

Annotate features in a Seurat object with additional metadata from databases or a GTF file.

Usage

AnnotateFeatures(
  srt,
  species = "Homo_sapiens",
  IDtype = c("symbol", "ensembl_id", "entrez_id"),
  db = NULL,
  db_update = FALSE,
  db_version = "latest",
  convert_species = TRUE,
  Ensembl_version = 103,
  mirror = NULL,
  gtf = NULL,
  merge_gtf_by = "gene_name",
  columns = c("seqname", "feature", "start", "end", "strand", "gene_id", "gene_name",
    "gene_type"),
  assays = "RNA",
  overwrite = FALSE
)

Arguments

srt

Seurat object to be annotated.

species

Name of the species to be used for annotation. Default is "Homo_sapiens".

IDtype

Type of identifier to use for annotation. Default is "symbol" with options "symbol", "ensembl_id", and "entrez_id".

db

Vector of database names to be used for annotation. Default is NULL.

db_update

Logical value indicating whether to update the database. Default is FALSE.

db_version

Version of the database to use. Default is "latest".

convert_species

A logical value indicating whether to use a species-converted database when the annotation is missing for the specified species. The default value is TRUE.

Ensembl_version

Version of the Ensembl database to use. Default is 103.

mirror

URL of the mirror to use for Ensembl database. Default is NULL.

gtf

Path to the GTF file to be used for annotation. Default is NULL.

merge_gtf_by

Column name to merge the GTF file by. Default is "gene_name".

columns

Vector of column names to be used from the GTF file. Default is "seqname", "feature", "start", "end", "strand", "gene_id", "gene_name", "gene_type".

assays

Character vector of assay names to be annotated. Default is "RNA".

overwrite

Logical value indicating whether to overwrite existing metadata. Default is FALSE.

See also

Examples

data(pancreas_sub)
pancreas_sub <- AnnotateFeatures(
  srt = pancreas_sub,
  species = "Mus_musculus",
  db = c(
    "Chromosome",
    "GeneType",
    "Enzyme",
    # "TF",
    # "CSPA",
    "VerSeDa"
  )
)
#> ℹ [2025-07-08 15:04:15] Species: Mus_musculus
#> ℹ [2025-07-08 15:04:16] Preparing database: Chromosome
#> ℹ [2025-07-08 15:04:17] Preparing database: GeneType
#> ℹ [2025-07-08 15:04:18] Preparing database: Enzyme
#> ℹ [2025-07-08 15:04:19] Preparing database: VerSeDa
#> ℹ [2025-07-08 15:04:24] Connect to the Ensembl archives...
#> ℹ [2025-07-08 15:04:25] Using the 103 version of biomart...
#> ℹ [2025-07-08 15:04:25] Connecting to the biomart...
#> ℹ [2025-07-08 15:04:27] Searching the dataset mmusculus ...
#> ℹ [2025-07-08 15:04:27] Connecting to the dataset mmusculus_gene_ensembl ...
#> ℹ [2025-07-08 15:04:28] Converting the geneIDs...
#> ℹ [2025-07-08 15:04:31] 11269 genes mapped with ensembl_peptide_id
#> ℹ [2025-07-08 15:04:33] ==============================
#>   [2025-07-08 15:04:33] 11269 genes mapped
#>   [2025-07-08 15:04:33] 1197 genes unmapped
#>   [2025-07-08 15:04:33] ==============================
#> ℹ [2025-07-08 15:04:37] Convert ID types for the database: Chromosome
#> ℹ [2025-07-08 15:04:37] Connect to the Ensembl archives...
#> ℹ [2025-07-08 15:04:38] Using the 103 version of biomart...
#> ℹ [2025-07-08 15:04:38] Connecting to the biomart...
#> ℹ [2025-07-08 15:04:38] Searching the dataset mmusculus ...
#> ℹ [2025-07-08 15:04:38] Connecting to the dataset mmusculus_gene_ensembl ...
#> ℹ [2025-07-08 15:04:39] Converting the geneIDs...
#> ℹ [2025-07-08 15:04:53] 27645 genes mapped with entrez_id
#> ℹ [2025-07-08 15:04:53] ==============================
#>   [2025-07-08 15:04:53] 27645 genes mapped
#>   [2025-07-08 15:04:53] 83811 genes unmapped
#>   [2025-07-08 15:04:53] ==============================
#> ℹ [2025-07-08 15:05:15] Convert ID types for the database: GeneType
#> ℹ [2025-07-08 15:05:15] Connect to the Ensembl archives...
#> ℹ [2025-07-08 15:05:15] Using the 103 version of biomart...
#> ℹ [2025-07-08 15:05:15] Connecting to the biomart...
#> ℹ [2025-07-08 15:05:16] Searching the dataset mmusculus ...
#> ℹ [2025-07-08 15:05:16] Connecting to the dataset mmusculus_gene_ensembl ...
#> ℹ [2025-07-08 15:05:17] Converting the geneIDs...
#> ℹ [2025-07-08 15:05:32] 27645 genes mapped with entrez_id
#> ℹ [2025-07-08 15:05:32] ==============================
#>   [2025-07-08 15:05:32] 27645 genes mapped
#>   [2025-07-08 15:05:32] 84660 genes unmapped
#>   [2025-07-08 15:05:32] ==============================
#> ℹ [2025-07-08 15:05:54] Convert ID types for the database: Enzyme
#> ℹ [2025-07-08 15:05:54] Connect to the Ensembl archives...
#> ℹ [2025-07-08 15:05:54] Using the 103 version of biomart...
#> ℹ [2025-07-08 15:05:54] Connecting to the biomart...
#> ℹ [2025-07-08 15:05:56] Searching the dataset mmusculus ...
#> ℹ [2025-07-08 15:05:56] Connecting to the dataset mmusculus_gene_ensembl ...
#> ℹ [2025-07-08 15:05:57] Converting the geneIDs...
#> ℹ [2025-07-08 15:05:57] 2263 genes mapped with entrez_id
#> ℹ [2025-07-08 15:05:57] ==============================
#>   [2025-07-08 15:05:57] 2263 genes mapped
#>   [2025-07-08 15:05:57] 20 genes unmapped
#>   [2025-07-08 15:05:57] ==============================
head(
  GetFeaturesData(
    pancreas_sub,
    assays = "RNA"
  )
)
#>                 TF CSPA Chromosome       GeneType Enzyme          VerSeDa
#> Xkr4          <NA> <NA>       chr1 protein-coding   <NA>             <NA>
#> Mrpl15        <NA> <NA>       chr1 protein-coding   <NA> SecretoryProtein
#> 4732440D04Rik <NA> <NA>       <NA>           <NA>   <NA>             <NA>
#> Gm26901       <NA> <NA>       <NA>           <NA>   <NA>             <NA>
#> Sntg1         <NA> <NA>       chr1 protein-coding   <NA> SecretoryProtein
#> Mybl1           TF <NA>       chr1 protein-coding   <NA>             <NA>

if (FALSE) { # \dontrun{
# Annotate features using a GTF file
pancreas_sub <- AnnotateFeatures(
  pancreas_sub,
  gtf = "/refdata-gex-mm10-2020-A/genes/genes.gtf"
)
head(
  GetFeaturesData(
    pancreas_sub,
    assays = "RNA"
  )
)
} # }