Prefetch cycle gene

Based on the human cell cycle genes, the cell cycle genes of the corresponding species were captured by homologous gene conversion.

Usage

CC_GenePrefetch(
  species = "Homo_sapiens",
  Ensembl_version = 103,
  mirror = NULL,
  max_tries = 5,
  use_cached_gene = TRUE,
  verbose = TRUE
)

Arguments

species

Latin names for animals,i.e., "Homo_sapiens", "Mus_musculus"

Ensembl_version

Ensembl database version. If NULL, use the current release version.

mirror

Specify an Ensembl mirror to connect to. The valid options here are 'www', 'uswest', 'useast', 'asia'.

max_tries

The maximum number of attempts to connect with the BioMart service.

use_cached_gene

Whether to use previously cached cell cycle gene conversion results for the species.

verbose

Whether to print messages.

Value

A list of S-phase and G2M-phase genes.

GeneConvert

Examples

ccgenes <- CC_GenePrefetch("Homo_sapiens")
#> ℹ [2025-07-26 06:24:23] Prefetching cell cycle genes forHomo_sapiens...
#> ✔ [2025-07-26 06:24:23] Cell cycle gene prefetching completed
str(ccgenes)
#> List of 3
#>  $ res: NULL
#>  $ S  : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ G2M: chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...
ccgenes <- CC_GenePrefetch("Mus_musculus")
#> ℹ [2025-07-26 06:24:23] Prefetching cell cycle genes forMus_musculus...
#> ℹ [2025-07-26 06:24:23] Connect to the Ensembl archives...
#> ℹ [2025-07-26 06:24:23] Using the 103 version of biomart...
#> ℹ [2025-07-26 06:24:23] Connecting to the biomart...
#> ℹ [2025-07-26 06:25:23] Error in `req_perform()`:
#> ℹ [2025-07-26 06:25:23] ! Failed to perform HTTP request.
#> ℹ [2025-07-26 06:25:23] Caused by error in `curl::curl_fetch_memory()`:
#> ℹ [2025-07-26 06:25:23] ! Timeout was reached [feb2021.archive.ensembl.org]:
#> ℹ [2025-07-26 06:25:23] Operation timed out after 60000 milliseconds with 0 bytes received
#> ℹ [2025-07-26 06:25:23] 
#> ℹ [2025-07-26 06:25:23] Get errors when connecting with ensembl mart...
#> ℹ [2025-07-26 06:25:24] Retrying...
#> ℹ [2025-07-26 06:26:24] Error in `req_perform()`:
#> ℹ [2025-07-26 06:26:24] ! Failed to perform HTTP request.
#> ℹ [2025-07-26 06:26:24] Caused by error in `curl::curl_fetch_memory()`:
#> ℹ [2025-07-26 06:26:24] ! Timeout was reached [feb2021.archive.ensembl.org]:
#> ℹ [2025-07-26 06:26:24] Operation timed out after 60001 milliseconds with 0 bytes received
#> ℹ [2025-07-26 06:26:24] 
#> ℹ [2025-07-26 06:26:24] Get errors when connecting with ensembl mart...
#> ℹ [2025-07-26 06:26:25] Retrying...
#> ℹ [2025-07-26 06:26:26] Searching the dataset hsapiens ...
#> ℹ [2025-07-26 06:26:26] Connecting to the dataset hsapiens_gene_ensembl ...
#> ℹ [2025-07-26 06:27:27] Error in `req_perform()`:
#> ℹ [2025-07-26 06:27:27] ! Failed to perform HTTP request.
#> ℹ [2025-07-26 06:27:27] Caused by error in `curl::curl_fetch_memory()`:
#> ℹ [2025-07-26 06:27:27] ! Timeout was reached [feb2021.archive.ensembl.org]:
#> ℹ [2025-07-26 06:27:27] Operation timed out after 60002 milliseconds with 0 bytes received
#> ℹ [2025-07-26 06:27:27] 
#> ℹ [2025-07-26 06:27:27] Get errors when connecting with Dataset(hsapiens_gene_ensembl)
#> ℹ [2025-07-26 06:27:28] Retrying...
#> ℹ [2025-07-26 06:28:28] Error in `req_perform()`:
#> ℹ [2025-07-26 06:28:28] ! Failed to perform HTTP request.
#> ℹ [2025-07-26 06:28:28] Caused by error in `curl::curl_fetch_memory()`:
#> ℹ [2025-07-26 06:28:28] ! Timeout was reached [feb2021.archive.ensembl.org]:
#> ℹ [2025-07-26 06:28:28] Operation timed out after 60000 milliseconds with 0 bytes received
#> ℹ [2025-07-26 06:28:28] 
#> ℹ [2025-07-26 06:28:28] Get errors when connecting with Dataset(hsapiens_gene_ensembl)
#> ℹ [2025-07-26 06:28:29] Retrying...
#> ℹ [2025-07-26 06:29:29] Error in `req_perform()`:
#> ℹ [2025-07-26 06:29:29] ! Failed to perform HTTP request.
#> ℹ [2025-07-26 06:29:29] Caused by error in `curl::curl_fetch_memory()`:
#> ℹ [2025-07-26 06:29:29] ! Timeout was reached [feb2021.archive.ensembl.org]:
#> ℹ [2025-07-26 06:29:29] Operation timed out after 60000 milliseconds with 0 bytes received
#> ℹ [2025-07-26 06:29:29] 
#> ℹ [2025-07-26 06:29:29] Get errors when connecting with Dataset(hsapiens_gene_ensembl)
#> ℹ [2025-07-26 06:29:30] Retrying...
#> ℹ [2025-07-26 06:30:30] Error in `req_perform()`:
#> ℹ [2025-07-26 06:30:30] ! Failed to perform HTTP request.
#> ℹ [2025-07-26 06:30:30] Caused by error in `curl::curl_fetch_memory()`:
#> ℹ [2025-07-26 06:30:30] ! Timeout was reached [feb2021.archive.ensembl.org]:
#> ℹ [2025-07-26 06:30:30] Operation timed out after 60000 milliseconds with 0 bytes received
#> ℹ [2025-07-26 06:30:30] 
#> ℹ [2025-07-26 06:30:30] Get errors when connecting with Dataset(hsapiens_gene_ensembl)
#> ℹ [2025-07-26 06:30:31] Retrying...
#> ℹ [2025-07-26 06:31:31] Error in `req_perform()`:
#> ℹ [2025-07-26 06:31:31] ! Failed to perform HTTP request.
#> ℹ [2025-07-26 06:31:31] Caused by error in `curl::curl_fetch_memory()`:
#> ℹ [2025-07-26 06:31:31] ! Timeout was reached [feb2021.archive.ensembl.org]:
#> ℹ [2025-07-26 06:31:31] Operation timed out after 60002 milliseconds with 0 bytes received
#> ℹ [2025-07-26 06:31:31] 
#> ℹ [2025-07-26 06:31:31] Get errors when connecting with Dataset(hsapiens_gene_ensembl)
#> Error in log_message(out, message_type = "error"): Error in `req_perform()`: ! Failed to perform HTTP request. Caused by
#> error in `curl::curl_fetch_memory()`: ! Timeout was reached
#> [feb2021.archive.ensembl.org]: Operation timed out after 60002 milliseconds
#> with 0 bytes received
str(ccgenes)
#> List of 3
#>  $ res: NULL
#>  $ S  : chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" ...
#>  $ G2M: chr [1:54] "HMGB2" "CDK1" "NUSAP1" "UBE2C" ...