Cluster tree plot

Visualize how cells move between clusters across multiple Seurat clustering resolutions.

Usage

ClusterTreePlot(
  srt,
  cluster_cols = NULL,
  prefix = NULL,
  resolutions = NULL,
  features = NULL,
  assay = NULL,
  layer = "data",
  edge_threshold = 0,
  node_palette = "Spectral",
  node_palcolor = NULL,
  edge_palette = "YlOrRd",
  edge_palcolor = NULL,
  node_size = c(3, 10),
  edge_size = c(0.25, 1.8),
  label = TRUE,
  label.size = 3,
  label.fg = "black",
  title = "Cluster tree",
  subtitle = NULL,
  xlab = "Resolution",
  ylab = "Cluster",
  legend.position = "bottom",
  theme_use = "theme_scop",
  theme_args = list(),
  combine = TRUE,
  ncol = NULL,
  return_data = FALSE,
  verbose = TRUE
)

Arguments

srt

A Seurat object.

cluster_cols

Metadata columns containing clustering results at different resolutions. If NULL, columns generated by Seurat::FindClusters(resolution = c(...)) are detected automatically.

prefix

Optional prefix used to filter automatically detected clustering columns, for example "RNA_snn".

resolutions

Optional numeric vector used to filter automatically detected resolutions.

features

A character vector of features to use.

assay

Which assay to use. If NULL, the default assay of the Seurat object will be used. When the object also contains ChromatinAssay, the default assay and additional ChromatinAssay will be preprocessed sequentially.

layer

Which layer to use. Default is data.

edge_threshold

Minimum incoming proportion required to draw an edge. Edges below this threshold are still available when return_data = TRUE.

node_palette, node_palcolor

Palette used for node colors.

edge_palette, edge_palcolor

Palette used for edge colors.

node_size

Numeric range used for node sizes.

edge_size

Numeric range used for edge widths.

label

Whether to add labels on the plot. Default is TRUE.

label.size

The size of the labels.

label.fg

The foreground color of the labels.

title

The title of the plot. Default is NULL.

subtitle

The subtitle of the plot. Default is NULL.

xlab

The label for the x-axis. Default is NULL.

ylab

The label for the y-axis. Default is NULL.

legend.position

The position of the legend. The default is "bottom" to keep the multiple cluster-tree guides compact.

theme_use

Theme used. Can be a character string or a theme function. Default is "theme_scop".

theme_args

Other arguments passed to the theme_use.

combine

Combine plots into a single patchwork object.

ncol

Number of columns used when combine = TRUE.

return_data

Whether to return plot data along with the plot.

verbose

Whether to show messages.

Value

A ggplot, patchwork, list of ggplot objects, or a list with plot data when return_data = TRUE.

Examples

if (FALSE) { # \dontrun{
data(pancreas_sub)
pancreas_sub <- standard_scop(pancreas_sub, verbose = FALSE)
pancreas_sub <- Seurat::FindNeighbors(
  pancreas_sub,
  dims = 1:20,
  verbose = FALSE
)
pancreas_sub <- Seurat::FindClusters(
  pancreas_sub,
  resolution = seq(0.2, 1, by = 0.2),
  verbose = FALSE
)
ClusterTreePlot(pancreas_sub)

ClusterTreePlot(
  pancreas_sub,
  features = c("Ins1", "Gcg")
)
} # }