Feature values on a 3D reduction

Feature values on a 3D reduction

Usage

FeatureDimPlot3D(
  srt,
  features,
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  split.by = NULL,
  layer = "data",
  assay = NULL,
  calculate_coexp = FALSE,
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE,
  verbose = TRUE
)

Arguments

srt

A Seurat object.

features

Features to plot: a character vector or a named list of assay gene names or numeric metadata columns.

reduction

Reduction to plot. NULL uses DefaultReduction.

dims

Length-2 vector of dimensions to plot.

axis_labs

Length-3 axis labels.

split.by

Metadata column to facet by.

layer

Assay layer to use.

assay

Assay to use. NULL uses the default assay.

calculate_coexp

Plot the geometric mean of the features.

shape.highlight

Shape of the cell to highlight. See scattergl-marker-symbol

width

Width in pixels, defaults to automatic sizing.

height

Height in pixels, defaults to automatic sizing.

save

The name of the file to save the plot to. Must end in ".html".

force

Draw even when more than 100 features are requested.

verbose

Print progress messages.

See also

Examples

data(pancreas_sub)
pancreas_sub <- RunStandardWorkflow(pancreas_sub)
#> ℹ [2026-08-21 15:59:03] Start standard processing workflow...
#> ℹ [2026-08-21 15:59:03] Checking a list of <Seurat>...
#> ! [2026-08-21 15:59:03] Data 1/1 of the `srt_list` is "unknown"
#> Warning: Data 1/1 of the `srt_list` is "unknown"
#> ℹ [2026-08-21 15:59:03] Perform `NormalizeData()` with `normalization.method = 'LogNormalize'` on 1/1 of `srt_list`...
#> ℹ [2026-08-21 15:59:03] Perform `FindVariableFeatures()` on 1/1 of `srt_list`...
#> ℹ [2026-08-21 15:59:03] Use the separate HVF from `srt_list`
#> ℹ [2026-08-21 15:59:03] Number of available HVF: 2000
#> ℹ [2026-08-21 15:59:03] Finished check
#> ℹ [2026-08-21 15:59:03] Perform `ScaleData()`
#> ℹ [2026-08-21 15:59:03] Perform pca linear dimension reduction
#> ℹ [2026-08-21 15:59:04] Use stored estimated dimensions 1:23 for Standardpca
#> ℹ [2026-08-21 15:59:04] Perform `Seurat::FindClusters()` with `cluster_algorithm = 'louvain'` and `cluster_resolution = 0.6`
#> ℹ [2026-08-21 15:59:04] Reorder clusters...
#> ℹ [2026-08-21 15:59:04] Skip `log1p()` because `layer = data` is not "counts"
#> ℹ [2026-08-21 15:59:04] Perform umap nonlinear dimension reduction
#> ✔ [2026-08-21 15:59:11] Standard processing workflow completed
FeatureDimPlot3D(
  pancreas_sub,
  features = c("Ghrl", "Ins1", "Gcg", "Ins2"),
  reduction = "StandardpcaUMAP3D"
)