3D-Dimensional reduction plot for gene expression visualization.

Plotting cell points on a reduced 3D space and coloring according to the gene expression in the cells.

Usage

FeatureDimPlot3D(
  srt,
  features,
  reduction = NULL,
  dims = c(1, 2, 3),
  axis_labs = NULL,
  split.by = NULL,
  layer = "data",
  assay = NULL,
  calculate_coexp = FALSE,
  pt.size = 1.5,
  cells.highlight = NULL,
  cols.highlight = "black",
  shape.highlight = "circle-open",
  sizes.highlight = 2,
  width = NULL,
  height = NULL,
  save = NULL,
  force = FALSE
)

Arguments

srt

A Seurat object.

features

A character vector or a named list of features to plot. Features can be gene names in Assay or names of numeric columns in meta.data.

reduction

Which dimensionality reduction to use. If not specified, will use the reduction returned by DefaultReduction.

dims

Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions.

axis_labs

A character vector of length 3 indicating the labels for the axes.

split.by

Name of a column in meta.data to split plot by.

layer

Which layer to pull expression data from? Default is data.

assay

Which assay to pull expression data from. If NULL, will use the assay returned by SeuratObject::DefaultAssay.

calculate_coexp

Whether to calculate the co-expression value (geometric mean) of the features.

pt.size

Point size for plotting.

cells.highlight

A vector of cell names to highlight.

cols.highlight

Color used to highlight the cells.

shape.highlight

Shape of the cell to highlight. See scattergl-marker-symbol

sizes.highlight

Size of highlighted cells.

width

Width in pixels, defaults to automatic sizing.

height

Height in pixels, defaults to automatic sizing.

save

The name of the file to save the plot to. Must end in ".html".

force

Whether to force drawing regardless of the number of features greater than 100.

See also

Examples

data(pancreas_sub)
pancreas_sub <- standard_scop(pancreas_sub)
#> ℹ [2025-08-17 10:21:19] Start scop standard workflow...
#> ℹ [2025-08-17 10:21:19] Checking srt_list...
#> ℹ [2025-08-17 10:21:20] Data is log-normalized
#> ℹ [2025-08-17 10:21:20] Data 1/1 of the `srt_list` has been log-normalized
#> ℹ [2025-08-17 10:21:20] Perform `Seurat::FindVariableFeatures()` on the data 1/1 of the `srt_list`...
#> ℹ [2025-08-17 10:21:20] Use the separate HVF from srt_list
#> ℹ [2025-08-17 10:21:21] Number of available HVF: 2000
#> ℹ [2025-08-17 10:21:21] Finished checking.
#> ℹ [2025-08-17 10:21:21] Perform `Seurat::ScaleData()` on the data...
#> ℹ [2025-08-17 10:21:21] Perform pca linear dimension reduction on the data...
#> ℹ [2025-08-17 10:21:21] `linear_reduction` pca is already existed. Skip calculation
#> ℹ [2025-08-17 10:21:21] Perform `Seurat::FindClusters()` with louvain and `cluster_resolution` = 0.6 on the data...
#> ℹ [2025-08-17 10:21:22] Reorder clusters...
#> ! [2025-08-17 10:21:22] Using `Seurat::AggregateExpression()` to calculate pseudo-bulk data for <Assay5>
#> ℹ [2025-08-17 10:21:22] Perform umap nonlinear dimension reduction on the data...
#> ℹ [2025-08-17 10:21:22] Non-linear dimensionality reduction (umap) using (Standardpca) dims (1-50) as input
#> ℹ [2025-08-17 10:21:25] Non-linear dimensionality reduction (umap) using (Standardpca) dims (1-50) as input
#> ✔ [2025-08-17 10:21:28] Run scop standard workflow done
FeatureDimPlot3D(
  pancreas_sub,
  features = c("Ghrl", "Ins1", "Gcg", "Ins2"),
  reduction = "StandardpcaUMAP3D"
)