List cached gene annotation databases

Lists the gene annotation databases cached by PrepareDB and PrepareCCCDB in the R.cache root, optionally filtered by species and database name. Each row describes one cached database with its version and creation date.

Usage

ListDB(species = c("Homo_sapiens", "Mus_musculus"), db = NULL)

Arguments

species

A character vector specifying the species. Can be "Homo_sapiens" or "Mus_musculus".

db

A character vector of database names (for example "GO_BP", "KEGG", "CellChat"), or a regular expression. If NULL, all databases are listed.

Value

A data frame with columns Database, Species, Version, and Date.

Examples

ListDB(species = "Homo_sapiens")
#> [1] Database Species  Version  Date    
#> <0 rows> (or 0-length row.names)
ListDB(species = c("Homo_sapiens", "Mus_musculus"))
#>   Database      Species                Version                       Date
#> 1     CSPA Mus_musculus           CSPA nterm:1 2026-08-15 13:59:12.130666
#> 2       DO Mus_musculus       9.0.0 nterm:6595 2026-08-15 14:01:58.959189
#> 3    GO_BP Mus_musculus     3.23.0 nterm:14957 2026-08-15 14:00:53.673134
#> 4    GO_CC Mus_musculus      3.23.0 nterm:2065 2026-08-15 14:00:54.883148
#> 5       MP Mus_musculus 2026-08-15 nterm:10916 2026-08-15 14:01:54.941521
#> 6       TF Mus_musculus    AnimalTFDB4 nterm:2 2026-08-15 13:32:04.359296
ListDB(species = "Mus_musculus", db = "GO_BP")
#>   Database      Species            Version                       Date
#> 1    GO_BP Mus_musculus 3.23.0 nterm:14957 2026-08-15 14:00:53.673134