Prepare the virtual environment

Prepare the virtual environment by installing the required dependencies and setting up the environment. This function prepares the virtual environment by checking if conda is installed, creating a new conda environment if needed, installing the required packages, and setting up the Python environment for use with scop. In order to create the environment, this function requires the path to the conda binary. If conda is set to "auto", it will attempt to automatically find the conda binary. If a conda environment with the specified name already exists and force is set to FALSE, the function will use the existing environment. If force set to TRUE, the existing environment will be recreated. Note that recreating the environment will remove any existing data in the environment. The function also checks if the package versions in the environment meet the requirements specified by the version parameter.

Usage

PrepareEnv(
  conda = "auto",
  miniconda_repo = "https://repo.anaconda.com/miniconda",
  envname = NULL,
  version = "3.10-1",
  force = FALSE,
  ...
)

Arguments

conda

The path to a conda executable. Use "auto" to allow scop to automatically find an appropriate conda binary.

miniconda_repo

Repositories for miniconda. Default is https://repo.anaconda.com/miniconda.

envname

The name of a conda environment.

version

A character vector specifying the version of the environment. Default is "3.10-1".

force

Whether to force a new environment to be created. If TRUE, the existing environment will be recreated. Default is FALSE.

...

Other arguments passed to reticulate::conda_install

Examples

PrepareEnv()
#> ℹ [2025-08-25 06:20:14] Preparing scop Python Environment
#> ℹ [2025-08-25 06:20:14] Environment name: scop_env
#> ℹ [2025-08-25 06:20:14] Python version: 3.10-1
#> ℹ [2025-08-25 06:20:14] Number of packages to install: 21
#> ℹ [2025-08-25 06:20:14] Auto-detecting conda...
#> ℹ [2025-08-25 06:20:14] Using existing environment: /usr/share/miniconda/envs/scop_env
#> ℹ [2025-08-25 06:20:14] Checking and installing required packages...
#> ℹ [2025-08-25 06:20:14] Installing conda packages
#> ℹ [2025-08-25 06:20:16] Checking 2 packages in environment: scop_env
#> ℹ [2025-08-25 06:20:16] Retrieving package list for environment: scop_env
#> ℹ [2025-08-25 06:20:18] Found 194 packages installed
#> ✔ [2025-08-25 06:20:18] leidenalg 0.10.2
#> ✔ [2025-08-25 06:20:18] tbb 2022.2.0
#> ℹ [2025-08-25 06:20:18] Installing pip packages
#> ℹ [2025-08-25 06:20:19] Checking 19 packages in environment: scop_env
#> ℹ [2025-08-25 06:20:20] Retrieving package list for environment: scop_env
#> ℹ [2025-08-25 06:20:22] Found 194 packages installed
#> ! [2025-08-25 06:20:22] igraph found but version mismatch: installed 0.10.16, required 0.11.9
#> ✔ [2025-08-25 06:20:22] matplotlib 3.10.3
#> ✔ [2025-08-25 06:20:22] numba 0.59.1
#> ✔ [2025-08-25 06:20:22] llvmlite 0.42.0
#> ✔ [2025-08-25 06:20:22] numpy 1.26.4
#> ✔ [2025-08-25 06:20:22] palantir 1.4.1
#> ✔ [2025-08-25 06:20:22] pandas 2.0.3
#> ✔ [2025-08-25 06:20:22] scanpy 1.11.3
#> ✔ [2025-08-25 06:20:22] scikit-learn 1.7.0
#> ✔ [2025-08-25 06:20:22] scipy 1.15.3
#> ✔ [2025-08-25 06:20:22] scvelo 0.3.3
#> ✔ [2025-08-25 06:20:22] wot 1.0.8.post2
#> ✔ [2025-08-25 06:20:22] trimap 1.1.4
#> ✔ [2025-08-25 06:20:22] pacmap 0.8.0
#> ✔ [2025-08-25 06:20:22] phate 1.0.11
#> ✔ [2025-08-25 06:20:22] bbknn 1.6.0
#> ✔ [2025-08-25 06:20:22] scanorama 1.7.4
#> ✔ [2025-08-25 06:20:22] scvi-tools 1.2.1
#> ✔ [2025-08-25 06:20:22] cellrank 2.0.7
#> ℹ [2025-08-25 06:20:22] Try to install: igraph==0.11.9
#> ℹ [2025-08-25 06:20:22] Installing 2 packages into environment: scop_env
#> ℹ [2025-08-25 06:20:23] Installing packages via pip...
#> ✔ [2025-08-25 06:20:25] pip installation completed
#> ℹ [2025-08-25 06:20:26] Checking 19 packages in environment: scop_env
#> ℹ [2025-08-25 06:20:27] Retrieving package list for environment: scop_env
#> ℹ [2025-08-25 06:20:29] Found 194 packages installed
#> ! [2025-08-25 06:20:29] igraph found but version mismatch: installed 0.10.16, required 0.11.9
#> ✔ [2025-08-25 06:20:29] matplotlib 3.10.3
#> ✔ [2025-08-25 06:20:29] numba 0.59.1
#> ✔ [2025-08-25 06:20:29] llvmlite 0.42.0
#> ✔ [2025-08-25 06:20:29] numpy 1.26.4
#> ✔ [2025-08-25 06:20:29] palantir 1.4.1
#> ✔ [2025-08-25 06:20:29] pandas 2.0.3
#> ✔ [2025-08-25 06:20:29] scanpy 1.11.3
#> ✔ [2025-08-25 06:20:29] scikit-learn 1.7.0
#> ✔ [2025-08-25 06:20:29] scipy 1.15.3
#> ✔ [2025-08-25 06:20:29] scvelo 0.3.3
#> ✔ [2025-08-25 06:20:29] wot 1.0.8.post2
#> ✔ [2025-08-25 06:20:29] trimap 1.1.4
#> ✔ [2025-08-25 06:20:29] pacmap 0.8.0
#> ✔ [2025-08-25 06:20:29] phate 1.0.11
#> ✔ [2025-08-25 06:20:29] bbknn 1.6.0
#> ✔ [2025-08-25 06:20:29] scanorama 1.7.4
#> ✔ [2025-08-25 06:20:29] scvi-tools 1.2.1
#> ✔ [2025-08-25 06:20:29] cellrank 2.0.7
#> ! [2025-08-25 06:20:29] Failed to install: igraph==0.11.9 into the environment scop_env. Please install manually
#> ✔ [2025-08-25 06:20:31] Python Environment Ready
#> conda environment: 
#>   conda:          /usr/share/miniconda/bin/conda
#>   environment:    /usr/share/miniconda/envs/scop_env
#> python config: 
#>   python:         /usr/share/miniconda/envs/scop_env/bin/python3.10
#>   libpython:      /usr/share/miniconda/envs/scop_env/lib/libpython3.10.so
#>   pythonhome:     /usr/share/miniconda/envs/scop_env:/usr/share/miniconda/envs/scop_env
#>   version:        3.10.1 | packaged by conda-forge | (main, Dec 22 2021, 01:39:36) [GCC 9.4.0]
#>   numpy:          /usr/share/miniconda/envs/scop_env/lib/python3.10/site-packages/numpy
#>   numpy_version:  1.26.4
#>   
#>   NOTE: Python version was forced by use_python() function