Run Harmony algorithm

This is a modified version of harmony::RunHarmony specifically designed for compatibility with RunSymphonyMap.

Usage

RunHarmony2(object, ...)

# S3 method for class 'Seurat'
RunHarmony2(
  object,
  group.by.vars,
  reduction = "pca",
  dims.use = 1:30,
  project.dim = TRUE,
  reduction.name = "Harmony",
  reduction.key = "Harmony_",
  verbose = TRUE,
  seed.use = 11L,
  ...
)

Arguments

object

A Seurat object.

...

Additional arguments to be passed to harmony::RunHarmony.

group.by.vars

The batch variable name.

reduction

The reduction to be used. Default is "pca".

dims.use

The dimensions to be used. Default is 1:30.

project.dim

Whether to project dimension reduction loadings. Default is TRUE.

reduction.name

The name of the reduction to be stored in the Seurat object. Default is "Harmony".

reduction.key

The prefix for the column names of the Harmony embeddings. Default is "Harmony_".

verbose

Whether to print the message. Default is TRUE.

seed.use

The random seed to be used. Default is 11.

Examples

panc8_sub <- RunHarmony2(
  panc8_sub,
  group.by.vars = "tech",
  reduction = "pca"
)
#> ℹ [2025-10-24 15:12:13] Installing: harmony...
#>  
#> → Will install 3 packages.
#> → All 3 packages (0 B) are cached.
#> + RcppProgress   0.4.2   
#> + RhpcBLASctl    0.23-42 
#> + harmony        1.1.0   [bld][cmp]
#>   
#> ℹ No downloads are needed, 3 pkgs are cached
#> ✔ Got harmony 1.1.0 (source) (5.19 MB)
#> ✔ Installed RcppProgress 0.4.2  (28ms)
#> ✔ Installed RhpcBLASctl 0.23-42  (37ms)
#> ℹ Building harmony 1.1.0
#> ✔ Built harmony 1.1.0 (27.1s)
#> ✔ Installed harmony 1.1.0  (100ms)
#> ✔ 1 pkg + 31 deps: kept 28, added 3, dld 1 (5.19 MB) [30.6s]
#> ℹ [2025-10-24 15:12:44] harmony@1.1.0 installed successfully
#> Error in object[[reduction]]: ‘pca’ not found in this Seurat object
#>  
CellDimPlot(
  panc8_sub,
  group.by = c("tech", "celltype"),
  reduction = "pca"
)
#> Error in DefaultReduction(srt, pattern = reduction): Unable to find any reductions
CellDimPlot(
  panc8_sub,
  group.by = c("tech", "celltype"),
  reduction = "Harmony"
)
#> Error in DefaultReduction(srt, pattern = reduction): Unable to find any reductions