Run Palantir analysis

Run Palantir analysis

Usage

RunPalantir(
  srt = NULL,
  assay_x = "RNA",
  layer_x = "counts",
  assay_y = c("spliced", "unspliced"),
  layer_y = "counts",
  adata = NULL,
  group_by = NULL,
  linear_reduction = NULL,
  nonlinear_reduction = NULL,
  basis = NULL,
  n_pcs = 30,
  n_neighbors = 30,
  dm_n_components = 10,
  dm_alpha = 0,
  dm_n_eigs = NULL,
  early_group = NULL,
  early_cell = NULL,
  terminal_cells = NULL,
  terminal_groups = NULL,
  num_waypoints = 1200,
  scale_components = TRUE,
  use_early_cell_as_start = TRUE,
  adjust_early_cell = FALSE,
  adjust_terminal_cells = FALSE,
  max_iterations = 25,
  n_jobs = 1,
  point_size = 20,
  palette = "Paired",
  palcolor = NULL,
  show_plot = TRUE,
  save = FALSE,
  dpi = 300,
  dirpath = "./",
  fileprefix = "",
  return_seurat = !is.null(srt)
)

Arguments

srt

A Seurat object.

assay_x

Assay to convert in the anndata object.

layer_x

Layer name for assay_x in the Seurat object.

assay_y

Assay to convert in the anndata object.

layer_y

Layer names for the assay_y in the Seurat object.

adata

An anndata object.

group_by

Variable to use for grouping cells in the Seurat object.

linear_reduction

Linear reduction method to use, e.g., "PCA".

nonlinear_reduction

Non-linear reduction method to use, e.g., "UMAP".

basis

The basis to use for reduction, e.g., "UMAP".

n_pcs

Number of principal components to use for linear reduction. Default is 30.

n_neighbors

Number of neighbors to use for constructing the KNN graph. Default is 30.

dm_n_components

The number of diffusion components to calculate.

dm_alpha

Normalization parameter for the diffusion operator.

dm_n_eigs

Number of eigen vectors to use.

early_group

Name of the group to start Palantir analysis from.

early_cell

Name of the cell to start Palantir analysis from.

terminal_cells

Character vector specifying terminal cells for Palantir analysis.

terminal_groups

Character vector specifying terminal groups for Palantir analysis.

num_waypoints

Number of waypoints to be included.

scale_components

Should the cell fate probabilities be scaled for each component independently?

use_early_cell_as_start

Should the starting cell for each terminal group be set as early_cell?

adjust_early_cell

Whether to adjust the early cell to the cell with the minimum pseudotime value.

adjust_terminal_cells

Whether to adjust the terminal cells to the cells with the maximum pseudotime value for each terminal group.

max_iterations

Maximum number of iterations for pseudotime convergence.

n_jobs

The number of parallel jobs to run.

point_size

The point size for plotting.

palette

The palette to use for coloring cells.

palcolor

A vector of colors to use as the palette.

show_plot

Whether to show the plot.

save

Whether to save the plots.

dpi

The DPI (dots per inch) for saving the plot.

dirpath

The directory to save the plots.

fileprefix

The file prefix to use for the plots.

return_seurat

Whether to return a Seurat object instead of an anndata object. Default is TRUE.

See also

Examples

PrepareEnv()
#> ℹ [2025-09-06 10:06:47] Preparing scop Python Environment
#> ℹ [2025-09-06 10:06:47] Environment name: scop_env
#> ℹ [2025-09-06 10:06:47] Python version: 3.10-1
#> ℹ [2025-09-06 10:06:47] Number of packages to install: 21
#> ℹ [2025-09-06 10:06:47] Auto-detecting conda...
#> ℹ [2025-09-06 10:06:48] Using existing environment: /usr/share/miniconda/envs/scop_env
#> ℹ [2025-09-06 10:06:48] Checking and installing required packages...
#> ℹ [2025-09-06 10:06:48] Installing conda packages
#> ℹ [2025-09-06 10:06:50] Checking 3 packages in environment: scop_env
#> ℹ [2025-09-06 10:06:50] Retrieving package list for environment: scop_env
#> ℹ [2025-09-06 10:06:51] Found 171 packages installed
#> ✔ [2025-09-06 10:06:51] leidenalg 0.10.2
#> ✔ [2025-09-06 10:06:51] tbb 2022.2.0
#> ✔ [2025-09-06 10:06:51] python-igraph 0.11.9
#> ℹ [2025-09-06 10:06:51] Installing pip packages
#> ℹ [2025-09-06 10:06:53] Checking 18 packages in environment: scop_env
#> ℹ [2025-09-06 10:06:53] Retrieving package list for environment: scop_env
#> ℹ [2025-09-06 10:06:54] Found 171 packages installed
#> ✔ [2025-09-06 10:06:55] matplotlib 3.10.3
#> ✔ [2025-09-06 10:06:55] numba 0.59.1
#> ✔ [2025-09-06 10:06:55] llvmlite 0.42.0
#> ✔ [2025-09-06 10:06:55] numpy 1.26.4
#> ! [2025-09-06 10:06:55] palantir not found
#> ✔ [2025-09-06 10:06:55] pandas 2.0.3
#> ! [2025-09-06 10:06:55] scanpy found but version mismatch: installed 1.11.4, required 1.11.3
#> ✔ [2025-09-06 10:06:55] scikit-learn 1.7.0
#> ✔ [2025-09-06 10:06:55] scipy 1.15.3
#> ✔ [2025-09-06 10:06:55] scvelo 0.3.3
#> ! [2025-09-06 10:06:55] wot not found
#> ✔ [2025-09-06 10:06:55] trimap 1.1.4
#> ✔ [2025-09-06 10:06:55] pacmap 0.8.0
#> ✔ [2025-09-06 10:06:55] phate 1.0.11
#> ! [2025-09-06 10:06:55] bbknn not found
#> ! [2025-09-06 10:06:55] scanorama not found
#> ! [2025-09-06 10:06:55] scvi-tools not found
#> ! [2025-09-06 10:06:55] cellrank not found
#> ℹ [2025-09-06 10:06:55] Try to install: palantir==1.4.1, scanpy==1.11.3, wot==1.0.8.post2, bbknn==1.6.0, scanorama==1.7.4, scvi-tools==1.2.1, and cellrank==2.0.7
#> ℹ [2025-09-06 10:06:55] Installing 8 packages into environment: scop_env
#> ℹ [2025-09-06 10:06:56] Installing packages via pip...
#> ℹ [2025-09-06 10:07:02] Checking 18 packages in environment: scop_env
#> ℹ [2025-09-06 10:07:03] Retrieving package list for environment: scop_env
#> ℹ [2025-09-06 10:07:04] Found 171 packages installed
#> ✔ [2025-09-06 10:07:04] matplotlib 3.10.3
#> ✔ [2025-09-06 10:07:04] numba 0.59.1
#> ✔ [2025-09-06 10:07:04] llvmlite 0.42.0
#> ✔ [2025-09-06 10:07:04] numpy 1.26.4
#> ! [2025-09-06 10:07:04] palantir not found
#> ✔ [2025-09-06 10:07:04] pandas 2.0.3
#> ! [2025-09-06 10:07:04] scanpy found but version mismatch: installed 1.11.4, required 1.11.3
#> ✔ [2025-09-06 10:07:04] scikit-learn 1.7.0
#> ✔ [2025-09-06 10:07:04] scipy 1.15.3
#> ✔ [2025-09-06 10:07:04] scvelo 0.3.3
#> ! [2025-09-06 10:07:04] wot not found
#> ✔ [2025-09-06 10:07:04] trimap 1.1.4
#> ✔ [2025-09-06 10:07:04] pacmap 0.8.0
#> ✔ [2025-09-06 10:07:04] phate 1.0.11
#> ! [2025-09-06 10:07:04] bbknn not found
#> ! [2025-09-06 10:07:04] scanorama not found
#> ! [2025-09-06 10:07:04] scvi-tools not found
#> ! [2025-09-06 10:07:04] cellrank not found
#> ! [2025-09-06 10:07:04] Failed to install: palantir==1.4.1, scanpy==1.11.3, wot==1.0.8.post2, bbknn==1.6.0, scanorama==1.7.4, scvi-tools==1.2.1, and cellrank==2.0.7 into the environment scop_env. Please install manually
#> ✔ [2025-09-06 10:07:07] Python Environment Ready
#> conda environment: 
#>   conda:          /usr/share/miniconda/bin/conda
#>   environment:    /usr/share/miniconda/envs/scop_env
#> python config: 
#>   python:         /usr/share/miniconda/envs/scop_env/bin/python3.10
#>   libpython:      /usr/share/miniconda/envs/scop_env/lib/libpython3.10.so
#>   pythonhome:     /usr/share/miniconda/envs/scop_env:/usr/share/miniconda/envs/scop_env
#>   version:        3.10.1 | packaged by conda-forge | (main, Dec 22 2021, 01:39:36) [GCC 9.4.0]
#>   numpy:          /usr/share/miniconda/envs/scop_env/lib/python3.10/site-packages/numpy
#>   numpy_version:  1.26.4
#>   
#>   NOTE: Python version was forced by use_python() function
data(pancreas_sub)
pancreas_sub <- RunPalantir(
  srt = pancreas_sub,
  group_by = "SubCellType",
  linear_reduction = "PCA",
  nonlinear_reduction = "UMAP",
  early_group = "Ductal",
  use_early_cell_as_start = TRUE,
  terminal_groups = c("Alpha", "Beta", "Delta", "Epsilon")
)
#> ℹ [2025-09-06 10:07:09] Checking 1 packages in environment: scop_env
#> ℹ [2025-09-06 10:07:09] Retrieving package list for environment: scop_env
#> ℹ [2025-09-06 10:07:10] Found 171 packages installed
#> ! [2025-09-06 10:07:10] palantir not found
#> ℹ [2025-09-06 10:07:10] Try to install: palantir
#> ℹ [2025-09-06 10:07:10] Installing 2 packages into environment: scop_env
#> ℹ [2025-09-06 10:07:12] Installing packages via pip...
#> ✔ [2025-09-06 10:07:15] pip installation completed
#> ℹ [2025-09-06 10:07:16] Checking 1 packages in environment: scop_env
#> ℹ [2025-09-06 10:07:16] Retrieving package list for environment: scop_env
#> ℹ [2025-09-06 10:07:17] Found 175 packages installed
#> ✔ [2025-09-06 10:07:17] palantir version: 1.4.1
#> ℹ [2025-09-06 10:07:20] Checking 2 packages in environment: scop_env
#> ℹ [2025-09-06 10:07:21] Retrieving package list for environment: scop_env
#> ℹ [2025-09-06 10:07:22] Found 175 packages installed
#> ✔ [2025-09-06 10:07:22] scanpy version: 1.11.4
#> ✔ [2025-09-06 10:07:22] numpy version: 1.26.4
#> ℹ [2025-09-06 10:07:22] Converting <Seurat> object to <AnnData> object...
#> ! [2025-09-06 10:07:22] "misc" slot is not converted
#> ! [2025-09-06 10:07:22] "tools" slot is not converted
#> ✔ [2025-09-06 10:07:22] Convert <Seurat> object to <AnnData> object completed
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named): TypeError: stack() got an unexpected keyword argument 'sort'
#> Run `reticulate::py_last_error()` for details.
head(pancreas_sub[[]])
#>                     orig.ident nCount_RNA nFeature_RNA     S_score   G2M_score
#> AAACCTGAGCCTTGAT SeuratProject       7071         2613 -0.01470664 -0.23261042
#> AAACCTGAGGCAATTA SeuratProject       4070         1984 -0.17125480 -0.28683382
#> AAACCTGTCTTTCCTC SeuratProject       9592         3410  0.55623430 -0.05274934
#> AAACGGGAGATATGGT SeuratProject       6194         2486 -0.16794573 -0.16668808
#> AAACGGGCAAGGACAC SeuratProject      10490         3281  0.21798572  0.68136787
#> AAACGGGCAGTTCCCT SeuratProject       6342         2430 -0.15680908 -0.16858186
#>                  nCount_spliced nFeature_spliced nCount_unspliced
#> AAACCTGAGCCTTGAT           7071             2613              959
#> AAACCTGAGGCAATTA           4070             1984             1076
#> AAACCTGTCTTTCCTC           9592             3410             1339
#> AAACGGGAGATATGGT           6194             2486              728
#> AAACGGGCAAGGACAC          10490             3281             1441
#> AAACGGGCAGTTCCCT           6342             2430             1113
#>                  nFeature_unspliced  CellType SubCellType Phase
#> AAACCTGAGCCTTGAT                643    Ductal      Ductal    G1
#> AAACCTGAGGCAATTA                701 Endocrine       Alpha    G1
#> AAACCTGTCTTTCCTC                894    Ductal      Ductal     S
#> AAACGGGAGATATGGT                556    Ductal      Ductal    G1
#> AAACGGGCAAGGACAC               1008    Ductal      Ductal   G2M
#> AAACGGGCAGTTCCCT                808    Ductal      Ductal    G1
FeatureDimPlot(
  pancreas_sub,
  c("palantir_pseudotime", "palantir_diff_potential")
)
#> ! [2025-09-06 10:07:24] Palantir_pseudotime,palantir_diff_potential are not in the features of srt.
#> Error in FeatureDimPlot(pancreas_sub, c("palantir_pseudotime", "palantir_diff_potential")): There are no valid features present.
FeatureDimPlot(
  pancreas_sub,
  paste0(
    c("Alpha", "Beta", "Delta", "Epsilon"),
    "_diff_potential"
  )
)
#> ! [2025-09-06 10:07:24] Alpha_diff_potential,Beta_diff_potential,Delta_diff_potential,Epsilon_diff_potential are not in the features of srt.
#> Error in FeatureDimPlot(pancreas_sub, paste0(c("Alpha", "Beta", "Delta",     "Epsilon"), "_diff_potential")): There are no valid features present.