Annotate single cells using SingleR

Annotate single cells using SingleR

RunSingleR(
  srt_query,
  srt_ref,
  query_group = NULL,
  ref_group = NULL,
  query_assay = "RNA",
  ref_assay = "RNA",
  genes = "de",
  de.method = "wilcox",
  sd.thresh = 1,
  de.n = NULL,
  aggr.ref = FALSE,
  aggr.args = list(),
  quantile = 0.8,
  fine.tune = TRUE,
  tune.thresh = 0.05,
  prune = TRUE,
  BPPARAM = BiocParallel::bpparam()
)

Arguments

srt_query

An object of class Seurat to be annotated with cell types.

srt_ref

An object of class Seurat storing the reference cells.

query_group

A character vector specifying the column name in the `srt_query` metadata that represents the cell grouping.

ref_group

A character vector specifying the column name in the `srt_ref` metadata that represents the cell grouping.

query_assay

A character vector specifying the assay to be used for the query data. Defaults to the default assay of the `srt_query` object.

ref_assay

A character vector specifying the assay to be used for the reference data. Defaults to the default assay of the `srt_ref` object.

genes

"genes" parameter in SingleR function.

de.method

"de.method" parameter in SingleR function.

sd.thresh

Deprecated and ignored.

de.n

An integer scalar specifying the number of DE genes to use when genes="de". If de.method="classic", defaults to 500 * (2/3) ^ log2(N) where N is the number of unique labels. Otherwise, defaults to 10. Ignored if genes is a list of markers/DE genes.

aggr.ref, aggr.args

Arguments controlling the aggregation of the references prior to annotation, see trainSingleR.

quantile, fine.tune, tune.thresh, prune

Further arguments to pass to classifySingleR.

BPPARAM

A BiocParallelParam object specifying how parallelization should be performed in other steps, see ?trainSingleR and ?classifySingleR for more details.

Examples

data("panc8_sub")
# Simply convert genes from human to mouse and preprocess the data
genenames <- make.unique(capitalize(rownames(panc8_sub), force_tolower = TRUE))
panc8_sub <- RenameFeatures(panc8_sub, newnames = genenames)
#> Rename features for the assay: RNA
panc8_sub <- check_srtMerge(panc8_sub, batch = "tech")[["srtMerge"]]
#> [2024-11-18 15:52:04.499911] Spliting srtMerge into srtList by column tech... ...
#> [2024-11-18 15:52:04.589138] Checking srtList... ...
#> Data 1/5 of the srtList is raw_normalized_counts. Perform NormalizeData(LogNormalize) on the data ...
#> Perform FindVariableFeatures on the data 1/5 of the srtList...
#> Data 2/5 of the srtList is raw_normalized_counts. Perform NormalizeData(LogNormalize) on the data ...
#> Perform FindVariableFeatures on the data 2/5 of the srtList...
#> Data 3/5 of the srtList is raw_normalized_counts. Perform NormalizeData(LogNormalize) on the data ...
#> Perform FindVariableFeatures on the data 3/5 of the srtList...
#> Data 4/5 of the srtList is raw_counts. Perform NormalizeData(LogNormalize) on the data ...
#> Perform FindVariableFeatures on the data 4/5 of the srtList...
#> Data 5/5 of the srtList is raw_counts. Perform NormalizeData(LogNormalize) on the data ...
#> Perform FindVariableFeatures on the data 5/5 of the srtList...
#> Use the separate HVF from srtList...
#> Number of available HVF: 2000
#> [2024-11-18 15:52:06.825973] Finished checking.

# Annotation
data("pancreas_sub")
pancreas_sub <- Standard_scop(pancreas_sub)
#> [2024-11-18 15:52:07.698341] Start Standard_scop
#> [2024-11-18 15:52:07.698427] Checking srtList... ...
#> Data 1/1 of the srtList is raw_counts. Perform NormalizeData(LogNormalize) on the data ...
#> Perform FindVariableFeatures on the data 1/1 of the srtList...
#> Use the separate HVF from srtList...
#> Number of available HVF: 2000
#> [2024-11-18 15:52:08.531195] Finished checking.
#> [2024-11-18 15:52:08.53126] Perform ScaleData on the data...
#> [2024-11-18 15:52:08.991335] Perform linear dimension reduction (pca) on the data...
#> Warning: The following arguments are not used: force.recalc
#> Warning: The following arguments are not used: force.recalc
#> [2024-11-18 15:52:10.185913] Perform FindClusters (louvain) on the data...
#> [2024-11-18 15:52:10.227584] Reorder clusters...
#> [2024-11-18 15:52:10.253448] Perform nonlinear dimension reduction (umap) on the data...
#> Non-linear dimensionality reduction(umap) using Reduction(Standardpca, dims:1-13) as input
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Non-linear dimensionality reduction(umap) using Reduction(Standardpca, dims:1-13) as input
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by ‘spam’
#> [2024-11-18 15:52:14.089415] Standard_scop done
#> Elapsed time: 6.39 secs 
pancreas_sub <- RunSingleR(
  srt_query = pancreas_sub, srt_ref = panc8_sub,
  query_group = "Standardclusters", ref_group = "celltype",
)
#> Detected srt_query data type: log_normalized_counts
#> Detected srt_ref data type: log_normalized_counts
#> Perform SingleRCluster on the data...
#>   |                                                                                                                      |                                                                                                              |   0%  |                                                                                                                      |===========                                                                                                   |  10%  |                                                                                                                      |======================                                                                                        |  20%  |                                                                                                                      |=================================                                                                             |  30%  |                                                                                                                      |============================================                                                                  |  40%  |                                                                                                                      |=======================================================                                                       |  50%  |                                                                                                                      |==================================================================                                            |  60%  |                                                                                                                      |=============================================================================                                 |  70%  |                                                                                                                      |========================================================================================                      |  80%  |                                                                                                                      |===================================================================================================           |  90%  |                                                                                                                      |==============================================================================================================| 100%
#> 
#>   |                                                                                                                      |                                                                                                              |   0%  |                                                                                                                      |=================                                                                                             |  15%  |                                                                                                                      |==================================                                                                            |  31%  |                                                                                                                      |===================================================                                                           |  46%  |                                                                                                                      |====================================================================                                          |  62%  |                                                                                                                      |============================================================================                                  |  69%  |                                                                                                                      |=============================================================================================                 |  85%  |                                                                                                                      |==============================================================================================================| 100%
#> 
#> Warning: missing values for 'group'
#>   |                                                                                                                      |                                                                                                              |   0%
#> 
#>   |                                                                                                                      |=======================================================                                                       |  50%
#> 
#>   |                                                                                                                      |==============================================================================================================| 100%
#> 
#> Error in `rownames<-`(`*tmp*`, value = new_rownames): missing values not allowed in rownames
CellDimPlot(pancreas_sub, group.by = "singler_annotation")
#> Error in CellDimPlot(pancreas_sub, group.by = "singler_annotation"): singler_annotation is not in the meta.data of srt object.

pancreas_sub <- RunSingleR(
  srt_query = pancreas_sub, srt_ref = panc8_sub,
  query_group = NULL, ref_group = "celltype"
)
#> Detected srt_query data type: log_normalized_counts
#> Detected srt_ref data type: log_normalized_counts
#> Perform SingleRCell on the data...
#>   |                                                                                                                      |                                                                                                              |   0%  |                                                                                                                      |===========                                                                                                   |  10%  |                                                                                                                      |======================                                                                                        |  20%  |                                                                                                                      |=================================                                                                             |  30%  |                                                                                                                      |============================================                                                                  |  40%  |                                                                                                                      |=======================================================                                                       |  50%  |                                                                                                                      |==================================================================                                            |  60%  |                                                                                                                      |=============================================================================                                 |  70%  |                                                                                                                      |========================================================================================                      |  80%  |                                                                                                                      |===================================================================================================           |  90%  |                                                                                                                      |==============================================================================================================| 100%
#> 
#>   |                                                                                                                      |                                                                                                              |   0%  |                                                                                                                      |=================                                                                                             |  15%  |                                                                                                                      |==================================                                                                            |  31%  |                                                                                                                      |===================================================                                                           |  46%  |                                                                                                                      |====================================================================                                          |  62%  |                                                                                                                      |============================================================================                                  |  69%  |                                                                                                                      |=============================================================================================                 |  85%  |                                                                                                                      |==============================================================================================================| 100%
#> 
#> Error: No cell overlap between new meta data and Seurat object
CellDimPlot(pancreas_sub, group.by = "singler_annotation")
#> Error in CellDimPlot(pancreas_sub, group.by = "singler_annotation"): singler_annotation is not in the meta.data of srt object.