Single-cell reference mapping with Symphony method

Single-cell reference mapping with Symphony method

RunSymphonyMap(
  srt_query,
  srt_ref,
  query_assay = NULL,
  ref_assay = srt_ref[[ref_pca]]@assay.used,
  ref_pca = NULL,
  ref_harmony = NULL,
  ref_umap = NULL,
  ref_group = NULL,
  projection_method = c("model", "knn"),
  nn_method = NULL,
  k = 30,
  distance_metric = "cosine",
  vote_fun = "mean"
)

Arguments

srt_query

An object of class Seurat storing the query cells.

srt_ref

An object of class Seurat storing the reference cells.

query_assay

A character string specifying the assay name for the query cells. If not provided, the default assay for the query object will be used.

ref_assay

A character string specifying the assay name for the reference cells. If not provided, the default assay for the reference object will be used.

ref_pca

A character string specifying the name of the PCA reduction in the reference object to use for calculating the distance metric.

ref_harmony

A character string specifying the name of the Harmony reduction in the reference object to use for calculating the distance metric.

ref_umap

A character string specifying the name of the UMAP reduction in the reference object. If not provided, the first UMAP reduction found in the reference object will be used.

ref_group

A character string specifying a metadata column name in the reference object to use for grouping.

projection_method

A character string specifying the projection method to use. Options are "model" and "knn". If "model" is selected, the function will try to use a pre-trained UMAP model in the reference object for projection. If "knn" is selected, the function will directly find the nearest neighbors using the distance metric.

nn_method

A character string specifying the nearest neighbor search method to use. Options are "raw", "annoy", and "rann". If "raw" is selected, the function will use the brute-force method to find the nearest neighbors. If "annoy" is selected, the function will use the Annoy library for approximate nearest neighbor search. If "rann" is selected, the function will use the RANN library for approximate nearest neighbor search. If not provided, the function will choose the search method based on the size of the query and reference datasets.

k

An integer specifying the number of nearest neighbors to find for each cell in the query object.

distance_metric

A character string specifying the distance metric to use for calculating the pairwise distances between cells. Options include: "pearson", "spearman", "cosine", "correlation", "jaccard", "ejaccard", "dice", "edice", "hamman", "simple matching", and "faith". Additional distance metrics can also be used, such as "euclidean", "manhattan", "hamming", etc.

vote_fun

A character string specifying the function to be used for aggregating the nearest neighbors in the reference object. Options are "mean", "median", "sum", "min", "max", "sd", "var", etc. If not provided, the default is "mean".

Examples

data("panc8_sub")
srt_ref <- panc8_sub[, panc8_sub$tech != "fluidigmc1"]
srt_query <- panc8_sub[, panc8_sub$tech == "fluidigmc1"]
srt_ref <- Integration_scop(srt_ref, batch = "tech", integration_method = "Harmony")
#> [2024-11-18 15:52:30.001799] Start Harmony_integrate
#> Error in packageVersion(pkg_name): there is no package called ‘harmony’
CellDimPlot(srt_ref, group.by = c("celltype", "tech"))
#> Error in DefaultReduction(srt): Unable to find any reductions.

# Projection
srt_query <- RunSymphonyMap(srt_query = srt_query, srt_ref = srt_ref, ref_pca = "Harmonypca", ref_harmony = "Harmony", ref_umap = "HarmonyUMAP2D")
#> Install package: "symphony" ...
#> 'getOption("repos")' replaces Bioconductor standard repositories, see 'help("repositories", package = "BiocManager")'
#> for details.
#> Replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> Bioconductor version 3.18 (BiocManager 1.30.23), R 4.3.2 (2023-10-31)
#> Installing github package(s) 'immunogenomics/symphony'
#> Warning: package ‘immunogenomics/symphony’ is not available for this version of R
#> 
#> A version of this package for your version of R might be available elsewhere,
#> see the ideas at
#> https://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages
#> Using GitHub PAT from the git credential store.
#> Downloading GitHub repo immunogenomics/symphony@HEAD
#> Installing 1 packages: harmony
#> Installing package into ‘/private/var/folders/rv/3jnfbs0d6r7d0c5bfj7ft5k00000gn/T/RtmpZ5Htsg/temp_libpath13d30c274401’
#> (as ‘lib’ is unspecified)
#> 
#> The downloaded binary packages are in
#> 	/var/folders/rv/3jnfbs0d6r7d0c5bfj7ft5k00000gn/T//RtmpAC8Cie/downloaded_packages
#> Adding ‘AUCell_1.24.0.tar.gz’ to the cache
#> Adding ‘GSEABase_1.64.0.tgz’ to the cache
#> Adding ‘RcppML_0.3.7.tgz’ to the cache
#> Adding ‘TTR_0.24.4.tgz’ to the cache
#> Adding ‘VIM_6.2.2.tgz’ to the cache
#> Adding ‘bluster_1.12.0.tgz’ to the cache
#> Adding ‘destiny_3.16.0.tgz’ to the cache
#> Adding ‘ggplot.multistats_1.0.1.tgz’ to the cache
#> Adding ‘glmGamPoi_1.14.3.tgz’ to the cache
#> Adding ‘glmpca_0.2.0.tgz’ to the cache
#> Adding ‘googleVis_0.7.3.tgz’ to the cache
#> Adding ‘graph_1.80.0.tgz’ to the cache
#> Adding ‘harmony_1.2.1.tgz’ to the cache
#> Adding ‘knn.covertree_1.0.tgz’ to the cache
#> Adding ‘laeken_0.5.3.tgz’ to the cache
#> Adding ‘metapod_1.10.1.tgz’ to the cache
#> Adding ‘pcaMethods_1.94.0.tgz’ to the cache
#> Adding ‘randomForest_4.7-1.2.tgz’ to the cache
#> Adding ‘ranger_0.17.0.tgz’ to the cache
#> Adding ‘scDblFinder_1.16.0.tgz’ to the cache
#> Adding ‘scater_1.30.1.tar.gz’ to the cache
#> Adding ‘scmap_1.24.0.tgz’ to the cache
#> Adding ‘scran_1.30.2.tgz’ to the cache
#> Adding ‘smoother_1.3.tgz’ to the cache
#> Adding ‘xgboost_1.7.8.1.tgz’ to the cache
#> Adding ‘xts_0.14.1.tgz’ to the cache
#> ── R CMD build ─────────────────────────────────────────────────────────────────────────────────────────────────────────
#> * checking for file ‘/private/var/folders/rv/3jnfbs0d6r7d0c5bfj7ft5k00000gn/T/RtmpAC8Cie/remotes48c01399248a/immunogenomics-symphony-bf7aaf9/DESCRIPTION’ ... OK
#> * preparing ‘symphony’:
#> * checking DESCRIPTION meta-information ... OK
#> * cleaning src
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘symphony_0.1.1.tar.gz’
#> 
#> Installing package into ‘/private/var/folders/rv/3jnfbs0d6r7d0c5bfj7ft5k00000gn/T/RtmpZ5Htsg/temp_libpath13d30c274401’
#> (as ‘lib’ is unspecified)
#> Adding ‘AUCell_1.24.0.tar.gz’ to the cache
#> Adding ‘GSEABase_1.64.0.tgz’ to the cache
#> Adding ‘RcppML_0.3.7.tgz’ to the cache
#> Adding ‘TTR_0.24.4.tgz’ to the cache
#> Adding ‘VIM_6.2.2.tgz’ to the cache
#> Adding ‘bluster_1.12.0.tgz’ to the cache
#> Adding ‘destiny_3.16.0.tgz’ to the cache
#> Adding ‘ggplot.multistats_1.0.1.tgz’ to the cache
#> Adding ‘glmGamPoi_1.14.3.tgz’ to the cache
#> Adding ‘glmpca_0.2.0.tgz’ to the cache
#> Adding ‘googleVis_0.7.3.tgz’ to the cache
#> Adding ‘graph_1.80.0.tgz’ to the cache
#> Adding ‘harmony_1.2.1.tgz’ to the cache
#> Adding ‘knn.covertree_1.0.tgz’ to the cache
#> Adding ‘laeken_0.5.3.tgz’ to the cache
#> Adding ‘metapod_1.10.1.tgz’ to the cache
#> Adding ‘pcaMethods_1.94.0.tgz’ to the cache
#> Adding ‘randomForest_4.7-1.2.tgz’ to the cache
#> Adding ‘ranger_0.17.0.tgz’ to the cache
#> Adding ‘scDblFinder_1.16.0.tgz’ to the cache
#> Adding ‘scater_1.30.1.tar.gz’ to the cache
#> Adding ‘scmap_1.24.0.tgz’ to the cache
#> Adding ‘scran_1.30.2.tgz’ to the cache
#> Adding ‘smoother_1.3.tgz’ to the cache
#> Adding ‘xgboost_1.7.8.1.tgz’ to the cache
#> Adding ‘xts_0.14.1.tgz’ to the cache
#> Error in srt_ref[[ref_pca]]: ‘Harmonypca’ not found in this Seurat object
#>  
ProjectionPlot(srt_query = srt_query, srt_ref = srt_ref, query_group = "celltype", ref_group = "celltype")
#> Error in srt_query[[query_reduction]]: ‘ref.embeddings’ not found in this Seurat object
#>