Convert an anndata object to a seurat object using reticulate

Convert an anndata object to a seurat object using reticulate

Usage

adata_to_srt(adata, verbose = TRUE)

Arguments

adata

A connected python anndata object.

verbose

Whether to print messages.

Examples

data(pancreas_sub)
pancreas_sub <- standard_scop(pancreas_sub)
#> ℹ [2025-09-12 11:14:43] Start standard scop workflow...
#> ℹ [2025-09-12 11:14:44] Checking a list of <Seurat> objects...
#> ! [2025-09-12 11:14:44] Data 1/1 of the `srt_list` is "unknown"
#> ℹ [2025-09-12 11:14:44] Perform `NormalizeData()` with `normalization.method = 'LogNormalize'` on the data 1/1 of the `srt_list`...
#> ℹ [2025-09-12 11:14:46] Perform `Seurat::FindVariableFeatures()` on the data 1/1 of the `srt_list`...
#> ℹ [2025-09-12 11:14:47] Use the separate HVF from srt_list
#> ℹ [2025-09-12 11:14:47] Number of available HVF: 2000
#> ℹ [2025-09-12 11:14:47] Finished check
#> ℹ [2025-09-12 11:14:47] Perform `Seurat::ScaleData()` on the data...
#> Warning: Different features in new layer data than already exists for scale.data
#> ℹ [2025-09-12 11:14:48] Perform pca linear dimension reduction on the data...
#> StandardPC_ 1 
#> Positive:  Aplp1, Cpe, Gnas, Fam183b, Map1b, Hmgn3, Pcsk1n, Chga, Tuba1a, Bex2 
#> 	   Syt13, Isl1, 1700086L19Rik, Pax6, Chgb, Scgn, Rbp4, Scg3, Gch1, Camk2n1 
#> 	   Cryba2, Pcsk2, Pyy, Tspan7, Mafb, Hist3h2ba, Dbpht2, Abcc8, Rap1b, Slc38a5 
#> Negative:  Spp1, Anxa2, Sparc, Dbi, 1700011H14Rik, Wfdc2, Gsta3, Adamts1, Clu, Mgst1 
#> 	   Bicc1, Ldha, Vim, Cldn3, Cyr61, Rps2, Mt1, Ptn, Phgdh, Nudt19 
#> 	   Smtnl2, Smco4, Habp2, Mt2, Col18a1, Rpl12, Galk1, Cldn10, Acot1, Ccnd1 
#> StandardPC_ 2 
#> Positive:  Rbp4, Tagln2, Tuba1b, Fkbp2, Pyy, Pcsk2, Iapp, Tmem27, Meis2, Tubb4b 
#> 	   Pcsk1n, Dbpht2, Rap1b, Dynll1, Tubb2a, Sdf2l1, Scgn, 1700086L19Rik, Scg2, Abcc8 
#> 	   Atp1b1, Hspa5, Fam183b, Papss2, Slc38a5, Scg3, Mageh1, Tspan7, Ppp1r1a, Ociad2 
#> Negative:  Neurog3, Btbd17, Gadd45a, Ppp1r14a, Neurod2, Sox4, Smarcd2, Mdk, Pax4, Btg2 
#> 	   Sult2b1, Hes6, Grasp, Igfbpl1, Gpx2, Cbfa2t3, Foxa3, Shf, Mfng, Tmsb4x 
#> 	   Amotl2, Gdpd1, Cdc14b, Epb42, Rcor2, Cotl1, Upk3bl, Rbfox3, Cldn6, Cer1 
#> StandardPC_ 3 
#> Positive:  Nusap1, Top2a, Birc5, Aurkb, Cdca8, Pbk, Mki67, Tpx2, Plk1, Ccnb1 
#> 	   2810417H13Rik, Incenp, Cenpf, Ccna2, Prc1, Racgap1, Cdk1, Aurka, Cdca3, Hmmr 
#> 	   Spc24, Kif23, Sgol1, Cenpe, Cdc20, Hist1h1b, Cdca2, Mxd3, Kif22, Ska1 
#> Negative:  Anxa5, Pdzk1ip1, Acot1, Tpm1, Anxa2, Dcdc2a, Capg, Sparc, Ttr, Pamr1 
#> 	   Clu, Cxcl12, Ndrg2, Hnf1aos1, Gas6, Gsta3, Krt18, Ces1d, Atp1b1, Muc1 
#> 	   Hhex, Acadm, Spp1, Enpp2, Bcl2l14, Sat1, Smtnl2, 1700011H14Rik, Tgm2, Fam159a 
#> StandardPC_ 4 
#> Positive:  Glud1, Tm4sf4, Akr1c19, Cldn4, Runx1t1, Fev, Pou3f4, Gm43861, Pgrmc1, Arx 
#> 	   Cd200, Lrpprc, Hmgn3, Ppp1r14c, Pam, Etv1, Tsc22d1, Slc25a5, Akap17b, Pgf 
#> 	   Fam43a, Emb, Jun, Krt8, Dnajc12, Mid1ip1, Ids, Rgs17, Uchl1, Alcam 
#> Negative:  Ins2, Ins1, Ppp1r1a, Nnat, Calr, Sytl4, Sdf2l1, Iapp, Pdia6, Mapt 
#> 	   G6pc2, C2cd4b, Npy, Gng12, P2ry1, Ero1lb, Adra2a, Papss2, Arhgap36, Fam151a 
#> 	   Dlk1, Creld2, Gip, Tmem215, Gm27033, Cntfr, Prss53, C2cd4a, Lyve1, Ociad2 
#> StandardPC_ 5 
#> Positive:  Pdx1, Nkx6-1, Npepl1, Cldn4, Cryba2, Fev, Jun, Chgb, Gng12, Adra2a 
#> 	   Mnx1, Sytl4, Pdk3, Gm27033, Nnat, Chga, Ins2, 1110012L19Rik, Enho, Krt7 
#> 	   Mlxipl, Tmsb10, Flrt1, Pax4, Tubb3, Prrg2, Gars, Frzb, BC023829, Gm2694 
#> Negative:  Irx2, Irx1, Gcg, Ctxn2, Tmem27, Ctsz, Tmsb15l, Nap1l5, Pou6f2, Gria2 
#> 	   Ghrl, Peg10, Smarca1, Arx, Lrpap1, Rgs4, Ttr, Gast, Tmsb15b2, Serpina1b 
#> 	   Slc16a10, Wnk3, Ly6e, Auts2, Sct, Arg1, Dusp10, Sphkap, Dock11, Edn3 
#> ℹ [2025-09-12 11:14:49] Perform `Seurat::FindClusters()` with louvain and `cluster_resolution` = 0.6 on the data...
#> ℹ [2025-09-12 11:14:49] Reorder clusters...
#> ! [2025-09-12 11:14:49] Using `Seurat::AggregateExpression()` to calculate pseudo-bulk data for <Assay5>
#> ℹ [2025-09-12 11:14:49] Perform umap nonlinear dimension reduction on the data...
#> ℹ [2025-09-12 11:14:49] Non-linear dimensionality reduction (umap) using (Standardpca) dims (1-50) as input
#> ℹ [2025-09-12 11:14:54] Non-linear dimensionality reduction (umap) using (Standardpca) dims (1-50) as input
#> ✔ [2025-09-12 11:14:58] Run scop standard workflow done
adata <- srt_to_adata(pancreas_sub)
#> ℹ [2025-09-12 11:15:00] Checking 2 packages in environment: scop_env
#> ℹ [2025-09-12 11:15:00] Retrieving package list for environment: scop_env
#> ℹ [2025-09-12 11:15:03] Found 194 packages installed
#> ✔ [2025-09-12 11:15:03] scanpy version: 1.11.3
#> ✔ [2025-09-12 11:15:03] numpy version: 1.26.4
#> ℹ [2025-09-12 11:15:03] Converting <Seurat> to <AnnData> ...
#> ! [2025-09-12 11:15:03] "misc" slot is not converted
#> ! [2025-09-12 11:15:03] "tools" slot is not converted
#> ✔ [2025-09-12 11:15:03] Convert <Seurat> object to <AnnData> object completed
adata <- RunPAGA(
  adata = adata,
  group_by = "SubCellType",
  linear_reduction = "PCA",
  nonlinear_reduction = "UMAP"
)
#> ℹ [2025-09-12 11:15:03] Running PAGA analysis...
#> Error in py_call_impl(callable, call_args$unnamed, call_args$named): ValueError: Did not find PCA in `.obsm.keys()`. You need to compute it first.
#> Run `reticulate::py_last_error()` for details.
srt <- adata_to_srt(adata)
#> ℹ [2025-09-12 11:15:03] Converting <AnnData> object to <Seurat> object...
#> ✔ [2025-09-12 11:15:04] Convert <AnnData> object to <Seurat> object completed
srt
#> An object of class Seurat 
#> 47994 features across 1000 samples within 3 assays 
#> Active assay: RNA (15998 features, 0 variable features)
#>  1 layer present: counts
#>  2 other assays present: spliced, unspliced
#>  9 dimensional reductions calculated: X_pca, X_X_pca, X_umap, X_X_umap, Standardpca, StandardpcaUMAP2D, StandardpcaUMAP3D, StandardUMAP2D, StandardUMAP3D

# Or convert a h5ad file to Seurat object
library(reticulate)
#> Error in library(reticulate): there is no package called ‘reticulate’
check_python("scanpy")
#> ℹ [2025-09-12 11:15:05] Checking 1 packages in environment: scop_env
#> ℹ [2025-09-12 11:15:06] Retrieving package list for environment: scop_env
#> ℹ [2025-09-12 11:15:08] Found 194 packages installed
#> ✔ [2025-09-12 11:15:08] scanpy version: 1.11.3
sc <- import("scanpy")
#> Error in import("scanpy"): could not find function "import"
adata <- sc$read_h5ad("pancreas.h5ad")
#> Error: object 'sc' not found
srt <- adata_to_srt(adata)
#> ℹ [2025-09-12 11:15:08] Converting <AnnData> object to <Seurat> object...
#> ✔ [2025-09-12 11:15:10] Convert <AnnData> object to <Seurat> object completed
srt
#> An object of class Seurat 
#> 47994 features across 1000 samples within 3 assays 
#> Active assay: RNA (15998 features, 0 variable features)
#>  1 layer present: counts
#>  2 other assays present: spliced, unspliced
#>  9 dimensional reductions calculated: X_pca, X_X_pca, X_umap, X_X_umap, Standardpca, StandardpcaUMAP2D, StandardpcaUMAP3D, StandardUMAP2D, StandardUMAP3D