Check and install R packages

Check and install R packages

Usage

check_r(packages, lib = .libPaths()[1], force = FALSE, verbose = TRUE)

Arguments

packages

Package to be installed. Package source can be CRAN, Bioconductor or Github. By default, the package name is extracted according to the packages parameter.

lib

The location of the library directories where to install the packages.

force

Whether to force the installation of packages. Default is FALSE.

verbose

Whether to print the message. Default is TRUE.

Examples

check_r(c("Seurat", "reticulate"))
#> ℹ [2025-10-22 10:08:43] Installing: Seurat...
#>  
#> → Will install 40 packages.
#> → All 40 packages (0 B) are cached.
#> + ROCR               1.0-11 
#> + RcppHNSW           0.6.0  
#> + RcppTOML           0.2.3  
#> + Seurat             5.3.0  
#> + SeuratObject       5.2.0  
#> + caTools            1.18.3 
#> + deldir             2.0-4  
#> + dotCall64          1.2    
#> + fastDummies        1.7.5  
#> + fitdistrplus       1.2-4  
#> + future             1.67.0 
#> + future.apply       1.20.0 
#> + globals            0.18.0 
#> + goftest            1.2-3  
#> + gplots             3.2.0  
#> + gtools             3.9.5  
#> + here               1.0.2  
#> + ica                1.0-3  
#> + leidenbase         0.1.35 
#> + listenv            0.9.1  
#> + lmtest             0.9-40 
#> + miniUI             0.1.2  
#> + parallelly         1.45.1 
#> + pbapply            1.7-4  
#> + progressr          0.17.0 
#> + reticulate         1.43.0  + ✔ python3
#> + rprojroot          2.1.1  
#> + scattermore        1.2    
#> + sctransform        0.4.2  
#> + sp                 2.2-0  
#> + spam               2.11-1 
#> + spatstat.data      3.1-9  
#> + spatstat.explore   3.5-3  
#> + spatstat.geom      3.6-0  
#> + spatstat.random    3.4-2  
#> + spatstat.sparse    3.1-0  
#> + spatstat.univar    3.1-4  
#> + spatstat.utils     3.2-0  
#> + tensor             1.5.1  
#> + zoo                1.8-14 
#> ✔ All system requirements are already installed.
#>   
#> ℹ No downloads are needed, 40 pkgs are cached
#> ℹ Installing system requirements
#> ℹ Executing `sudo sh -c apt-get -y update`
#> Get:1 file:/etc/apt/apt-mirrors.txt Mirrorlist [144 B]
#> Hit:2 http://azure.archive.ubuntu.com/ubuntu noble InRelease
#> Hit:3 http://azure.archive.ubuntu.com/ubuntu noble-updates InRelease
#> Hit:4 http://azure.archive.ubuntu.com/ubuntu noble-backports InRelease
#> Hit:5 http://azure.archive.ubuntu.com/ubuntu noble-security InRelease
#> Hit:6 https://packages.microsoft.com/repos/azure-cli noble InRelease
#> Hit:7 https://packages.microsoft.com/ubuntu/24.04/prod noble InRelease
#> Reading package lists...
#> ℹ Executing `sudo sh -c apt-get -y install python3 libcurl4-openssl-dev libssl-dev make zlib1g-dev libglpk-dev libxml2-dev pandoc libpng-dev libicu-dev`
#> Reading package lists...
#> Building dependency tree...
#> Reading state information...
#> python3 is already the newest version (3.12.3-0ubuntu2).
#> libcurl4-openssl-dev is already the newest version (8.5.0-2ubuntu10.6).
#> libssl-dev is already the newest version (3.0.13-0ubuntu3.6).
#> make is already the newest version (4.3-4.1build2).
#> zlib1g-dev is already the newest version (1:1.3.dfsg-3.1ubuntu2.1).
#> libglpk-dev is already the newest version (5.0-1build2).
#> libxml2-dev is already the newest version (2.9.14+dfsg-1.3ubuntu3.5).
#> pandoc is already the newest version (3.1.3+ds-2).
#> libpng-dev is already the newest version (1.6.43-5build1).
#> libicu-dev is already the newest version (74.2-1ubuntu3.1).
#> 0 upgraded, 0 newly installed, 0 to remove and 36 not upgraded.
#> ✔ Installed caTools 1.18.3  (97ms)
#> ✔ Installed deldir 2.0-4  (113ms)
#> ✔ Installed dotCall64 1.2  (99ms)
#> ✔ Installed fastDummies 1.7.5  (135ms)
#> ✔ Installed fitdistrplus 1.2-4  (70ms)
#> ✔ Installed future 1.67.0  (72ms)
#> ✔ Installed future.apply 1.20.0  (71ms)
#> ✔ Installed globals 0.18.0  (99ms)
#> ✔ Installed goftest 1.2-3  (64ms)
#> ✔ Installed gplots 3.2.0  (67ms)
#> ✔ Installed gtools 3.9.5  (69ms)
#> ✔ Installed here 1.0.2  (70ms)
#> ✔ Installed ica 1.0-3  (65ms)
#> ✔ Installed leidenbase 0.1.35  (96ms)
#> ✔ Installed listenv 0.9.1  (70ms)
#> ✔ Installed lmtest 0.9-40  (66ms)
#> ✔ Installed miniUI 0.1.2  (65ms)
#> ✔ Installed parallelly 1.45.1  (67ms)
#> ✔ Installed pbapply 1.7-4  (69ms)
#> ✔ Installed progressr 0.17.0  (66ms)
#> ✔ Installed RcppHNSW 0.6.0  (96ms)
#> ✔ Installed RcppTOML 0.2.3  (66ms)
#> ✔ Installed reticulate 1.43.0  (71ms)
#> ✔ Installed ROCR 1.0-11  (71ms)
#> ✔ Installed rprojroot 2.1.1  (71ms)
#> ✔ Installed scattermore 1.2  (67ms)
#> ✔ Installed sctransform 0.4.2  (95ms)
#> ✔ Installed Seurat 5.3.0  (73ms)
#> ✔ Installed SeuratObject 5.2.0  (76ms)
#> ✔ Installed sp 2.2-0  (1.1s)
#> ✔ Installed spam 2.11-1  (1.1s)
#> ✔ Installed spatstat.data 3.1-9  (80ms)
#> ✔ Installed spatstat.explore 3.5-3  (78ms)
#> ✔ Installed spatstat.geom 3.6-0  (73ms)
#> ✔ Installed spatstat.random 3.4-2  (69ms)
#> ✔ Installed spatstat.sparse 3.1-0  (98ms)
#> ✔ Installed spatstat.univar 3.1-4  (66ms)
#> ✔ Installed spatstat.utils 3.2-0  (66ms)
#> ✔ Installed tensor 1.5.1  (63ms)
#> ✔ Installed zoo 1.8-14  (46ms)
#> ✔ 1 pkg + 134 deps: kept 93, added 40 [9.2s]
#> ℹ [2025-10-22 10:08:52] Seurat and reticulate installed successfully