Convert a Seurat object to an AnnData object

This function takes a Seurat object and converts it to an anndata object using the reticulate package.

Usage

srt_to_adata(
  srt,
  features = NULL,
  assay_x = "RNA",
  layer_x = "counts",
  assay_y = c("spliced", "unspliced"),
  layer_y = "counts",
  convert_tools = FALSE,
  convert_misc = FALSE,
  verbose = TRUE
)

Arguments

srt

A Seurat object.

features

Optional vector of features to include in the anndata object. Defaults to all features in assay_x.

assay_x

Assay to convert as the main data matrix in the anndata object.

layer_x

Layer name for assay_x in the Seurat object.

assay_y

Assays to convert as layers in the anndata object.

layer_y

Layer names for the assay_y in the Seurat object.

convert_tools

Whether to convert the tool-specific data. Default is FALSE.

convert_misc

Whether to convert the miscellaneous data. Default is FALSE.

verbose

Whether to print verbose messages during the conversion process. Default is TRUE.

Value

A anndata object.

Examples

data(pancreas_sub)
adata <- srt_to_adata(pancreas_sub)
#> ℹ [2025-08-25 06:57:06] Checking 2 packages in environment: scop_env
#> ℹ [2025-08-25 06:57:06] Retrieving package list for environment: scop_env
#> ℹ [2025-08-25 06:57:08] Found 194 packages installed
#> ✔ [2025-08-25 06:57:08] scanpy version: 1.11.3
#> ✔ [2025-08-25 06:57:08] numpy version: 1.26.4
#> ℹ [2025-08-25 06:57:10] Converting <Seurat> object to <AnnData> object...
#> ! [2025-08-25 06:57:11] "misc" slot is not converted
#> ! [2025-08-25 06:57:11] "tools" slot is not converted
#> ✔ [2025-08-25 06:57:11] Convert <Seurat> object to <AnnData> object completed
adata
#> AnnData object with n_obs × n_vars = 1000 × 15962
#>     obs: 'orig.ident', 'nCount_RNA', 'nFeature_RNA', 'S_score', 'G2M_score', 'nCount_spliced', 'nFeature_spliced', 'nCount_unspliced', 'nFeature_unspliced', 'CellType', 'SubCellType', 'Phase'
#>     var: 'features', 'TF', 'CSPA', 'highly_variable'
#>     obsm: 'pca', 'X_pca', 'umap', 'X_umap'
#>     layers: 'spliced', 'unspliced'
#>     obsp: 'distances', 'connectivities'

if (FALSE) { # \dontrun{
# Or save as a h5ad/loom file
adata$write_h5ad(
  "pancreas_sub.h5ad"
)
adata$write_loom(
  "pancreas_sub.loom",
  write_obsm_varm = TRUE
)
} # }